### 2023-01-07 #5055 @danielskatz
👋 @pitmonticone - thanks for your submission. I'll now look for an editor.

While I do so, note that the paper is quite long (~2300 words compared to a suggested maximum of 1000). I suggest some of the text might be removed in favor of pointers to the documentation or README.  Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. You can also use `@editorialbot check repository` to check the length of the paper. editorialbot commands need to be the first entry in a new comment.
