### 2023-01-07 #5055 @pitmonticone
Hi @danielskatz, thank you very much for your ready response with helpful feedbacks and clear instructions.

Here are some potential reviewers extracted from your list:
- juliohm: "Julia" among preferred languages and "graph theory" among preferred domains;
- tpoisot: "Julia" among preferred languages and "graph theory" among preferred domains;
- Datseris: "Julia" among preferred languages and "complex systems" among preferred domains;
- bgailleton: "Julia" among other languages and "graph theory" among preferred domains;
- TomKellyGenetics: "Julia" among other languages and "graph theory" among preferred domains;
- amritagos: "Julia" among other languages and "graph theory" among preferred domains.

We (@ClaudMor and I) are going to shorten the [paper](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper.md) as much as possible as soon as we can.

Thanks again for your work.
