+++ paper/tmp.md
@@ -55,7 +55,7 @@ Several theoretical frameworks have been proposed to formally subsume all instan
 
 Multilayer graphs have been adopted to model the structure and dynamics of a wide spectrum of high-dimensional, non-linear, multi-scale, time-dependent complex systems including physical, chemical, biological, neuronal, socio-technical, epidemiological, ecological and economic networks [@Cozzo2013; @Granell2013; @Massaro2014; @Estrada2014; @AzimiTafreshi2016; @Baggio2016; @DeDomenico2016; @Amato2017; @DeDomenico2017; @Pilosof2017; @deArruda2017; @Gosak2018; @SorianoPaos2018; @Timteo2018; @Buldu2018; @Lim2019; @Mangioni2020; @Aleta2020; @Aleta2022]. 
 
-We have chosen the [Julia language](https://julialang.org) for this software package because it is a modern, open-source, high-level, high-performance dynamic language for technical computing [@Bezanson2017]. At the best of our knowledge there are currently no software packages dedicated to the creation, manipulation and analysis of multilayer graphs implemented in the Julia language apart from MultilayerGraphs.jl itself [@Moroni_Monticone_MultilayerGraphs_2022]. 
+At the best of our knowledge there are currently no software packages dedicated to the creation, manipulation and analysis of multilayer graphs implemented in the Julia language [@Bezanson2017] apart from MultilayerGraphs.jl itself [@Moroni_Monticone_MultilayerGraphs_2022]. 
 
 # Main Features 
 
@@ -83,60 +83,49 @@ To install MultilayerGraphs.jl it is sufficient to activate the `pkg` mode by pr
 pkg> add MultilayerGraphs
 ```
 
-In the following code chunks we synthetically illustrate some of the main features outlined in the previous section. 
-
-Let's begin by importing the necessary dependencies and setting the relevant constants.
+In the following code chunks we synthetically illustrate: 
+- how to define **layers** and **interlayers** with a variety of constructors and underlying graphs;
+- how to construct **directed multilayer graph** with those layers and interlayers;
+- how to add nodes, vertices and edges to the multilayer graph;
+- how to compute some multilayer metrics as defined in @DeDomenico2013. 
 
 ```julia
+# Import necessary dependencies
 using Distributions, Graphs, SimpleValueGraphs
 using MultilayerGraphs
 
-# Set the number of nodes: objects represented by multilayer vertices
-const n_nodes = 100 
-# Create a list of nodes
-const node_list = [Node("node_$i") for i in 1:n_nodes]
-```
+# Set relevant constants 
+n_nodes = 100 
+node_list = [Node("node_$i") for i in 1:n_nodes]
 
-## Layers and Interlayers 
-
-We will instantiate layers and interlayers with randomly-selected edges and vertices adopting a variety of techniques. Layers and Interlayers are not immutable, and mostly behave like normal graphs. The reader is invited to consult the [API](https://juliagraphs.org/MultilayerGraphs.jl/stable/API/) for more information.
-
-Here we define a layer with an underlying simple directed graph using a graph generator-like (or "configuration model"-like) constructor which allows us to specify both the **indegree** and the **outdegree sequences**. Before instantiating each layer we sample the number of its vertices and, optionally, of its edges.
-
-```julia
-n_vertices = rand(1:100)                          # Number of vertices 
-layer_simple_directed = layer_simpledigraph(      # Layer constructor 
-    :layer_simple_directed,                       # Layer name
-    sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
-    Truncated(Normal(5, 5), 0, 20), # Indegree sequence distribution 
-    Truncated(Normal(5, 5), 0, 20)  # Outdegree sequence distribution
+# Simple directed layer 
[…]
