+++ paper/paper_long.md
@@ -1,0 +1,288 @@
+---
+title: 'MultilayerGraphs.jl: Multilayer Network Science in Julia'
+tags:
+  - Julia
+  - Graphs
+  - Networks
+  - Complexity
+  - Complex-Networks
+  - Complex-Systems
+  - Network-Science
+  - Network-Analysis
+  - Multilayer-Graphs
+  - Multilayer-Networks
+authors:
+  - name: Claudio Moroni
+    orcid: 0000-0003-1274-6937
+    equal-contrib: true
+    affiliation: "1, 2" 
+  - name: Pietro Monticone
+    orcid: 0000-0002-2731-9623
+    equal-contrib: true
+    affiliation: "1, 2" 
+affiliations:
+ - name: University of Turin, Italy
+   index: 1
+ - name: Interdisciplinary Physics Team, Italy
+   index: 2
+date: 30 December 2022
+bibliography: paper.bib
+---
+
+# Summary
+
+**MultilayerGraphs.jl** is a Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs.
+
+A multilayer graph consists of multiple subgraphs called *layers* which can be interconnected through [bipartite graphs](https://en.wikipedia.org/wiki/Bipartite_graph) called *interlayers* composed of the vertex sets of two different layers and the edges between them. The vertices in each layer represent a single set of nodes, although not all nodes have to be represented in every layer. 
+
+Formally, a multilayer graph can be defined as a triple $G=(V,E,L)$, where:
+
+- $V$ is the set of vertices;
+- $E$ is the set of edges, pairs of nodes $(u, v)$ representing a connection, relationship or interaction between the nodes $u$ and $v$;
+- $L$ is a set of layers, which are subsets of $V$ and $E$ encoding the nodes and edges within each layer.
+
+Each layer $\ell$ in $L$ is a tuple $(V_\ell, E_\ell)$, where $V_\ell$ is a subset of $V$ that represents the vertices within that layer, and $E_\ell$ is a subset of $E$ that represents the edges within that layer.
+
+MultilayerGraphs.jl is an integral part of the [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem extending Graphs.jl [@Graphs2021] so all the methods and metrics exported by Graphs.jl work for multilayer graphs, but due to the special nature of multilayer graphs the package features a peculiar implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects an experienced practitioner would expect such as nodes (`Node`), vertices (`MultilayerVertex`), layers (`Layer`), interlayers (`Interlayer`), etc.
+
+MultilayerGraphs.jl features multilayer-specific methods and metrics including the global clustering coefficient, the overlay clustering coefficient, the multilayer eigenvector centrality, the multilayer modularity and the Von Neumann entropy.
+
+Finally, MultilayerGraphs.jl has been integrated within the [JuliaDynamics](https://github.com/JuliaDynamics) ecosystem so that any `Multilayer(Di)Graph` can be utilised as an argument to the `GraphSpace` constructor in Agents.jl [@Datseris2022]. 
+
+# Statement of Need
+
+Several theoretical frameworks have been proposed to formally subsume all instances of multilayer graphs [@DeDomenico2013; @Kivela2014; @Boccaletti2014; @Lee2015; @Aleta2019; @Bianconi2018; @Cozzo2018; @Artime2022; @DeDomenico2022]. 
+
[…]
