+++ .github/workflows/PaperDraft.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
-          name: paper
+          name: paper_short
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
