+++ .github/workflows/PaperDraft.yml
@@ -16,8 +16,8 @@ jobs:
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
-          name: paper_short
+          name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper/paper_short.pdf
+          path: paper
