+++ .github/workflows/PaperDraft.yml
@@ -12,12 +12,12 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper/paper.md
+          paper-path: paper/paper_short.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
-          name: paper
+          name: paper_short
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper/paper.pdf
+          path: paper/paper_short.pdf
