+++ paper/paper_short.md
@@ -84,6 +84,7 @@ pkg> add MultilayerGraphs
 ```
 
 In the following code chunk we synthetically illustrate: 
+
 - how to define **layers** and **interlayers** with a variety of constructors and underlying graphs;
 - how to construct **directed multilayer graph** with those layers and interlayers;
 - how to add nodes, vertices and edges to the multilayer graph;
