+++ paper/paper.md
@@ -86,7 +86,7 @@ pkg> add MultilayerGraphs
 In the following code chunk we synthetically illustrate: 
 
 - how to define **layers** and **interlayers** with a variety of constructors and underlying graphs;
-- how to construct **directed multilayer graph** with those layers and interlayers;
+- how to construct a **directed multilayer graph** with those layers and interlayers;
 - how to add nodes, vertices and edges to the multilayer graph;
 - how to compute some multilayer metrics as defined in @DeDomenico2013. 
 
