+++ paper/paper_medium.md
@@ -43,9 +43,9 @@ Formally, a multilayer graph can be defined as a triple $G=(V,E,L)$, where:
 
 Each layer $\ell$ in $L$ is a tuple $(V_\ell, E_\ell)$, where $V_\ell$ is a subset of $V$ that represents the vertices within that layer, and $E_\ell$ is a subset of $E$ that represents the edges within that layer.
 
-MultilayerGraphs.jl is an integral part of the [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem extending Graphs.jl [@Graphs2021] so all the methods and metrics exported by Graphs.jl work for multilayer graphs, but due to the special nature of multilayer graphs the package features a peculiar implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects an experienced practitioner would expect such as nodes (`Node`), vertices (`MultilayerVertex`), layers (`Layer`), interlayers (`Interlayer`), etc.
+MultilayerGraphs.jl is an integral part of the [JuliaGraphs](https://github.com/JuliaGraphs) ecosystem extending Graphs.jl [@Graphs2021] so all the methods and metrics exported by Graphs.jl work for multilayer graphs, but due to the special nature of multilayer graphs the package features a peculiar implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects an experienced practitioner would expect such as nodes, vertices, layers, interlayers, etc.
 
-MultilayerGraphs.jl features multilayer-specific methods and metrics including the global clustering coefficient, the overlay clustering coefficient, the multilayer eigenvector centrality, the multilayer modularity and the Von Neumann entropy.
+MultilayerGraphs.jl features specific methods and metrics including the global clustering coefficient, the overlay clustering coefficient, the multilayer eigenvector centrality, the multilayer modularity and the Von Neumann entropy.
 
 Finally, MultilayerGraphs.jl has been integrated within the [JuliaDynamics](https://github.com/JuliaDynamics) ecosystem so that any `Multilayer(Di)Graph` can be utilised as an argument to the `GraphSpace` constructor in Agents.jl [@Datseris2022]. 
 
@@ -75,11 +75,160 @@ Alternatively, it is possible to construct a `Multilayer(Di)Graph` making use of
 
 For a more comprehensive exploration of the package features and functionalities we strongly recommend consulting the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs.jl).  
 
+# Installation and Usage 
+
+To install MultilayerGraphs.jl it is sufficient to activate the `pkg` mode by pressing `]` in the Julia REPL and then run the following command:
+
+```nothing
+pkg> add MultilayerGraphs
+```
+
+In the following code chunk we synthetically illustrate: 
+
+- how to define **layers** and **interlayers** with a variety of constructors and underlying graphs;
+- how to construct a **directed multilayer graph** with those layers and interlayers;
+- how to add nodes, vertices and edges to the multilayer graph;
+- how to compute some multilayer metrics as defined in @DeDomenico2013. 
+
+```julia
+# Import necessary dependencies
+using Distributions, Graphs, SimpleValueGraphs
+using MultilayerGraphs
+
+# Set relevant constants 
+n_nodes = 100 
+node_list = [Node("node_$i") for i in 1:n_nodes]
+
+# Simple directed layer 
+n_vertices = rand(1:100)                          
+layer_simple_directed = layer_simpledigraph(      
+    :layer_simple_directed,                       
+    sample(node_list, n_vertices; replace=false), 
[…]
