### 2023-01-08 #5055 @diehlpk
@danielskatz Sure, I am traveling until Thursday and will start after that.
