+++ paper/paper.md
@@ -75,11 +75,203 @@ Alternatively, it is possible to construct a `Multilayer(Di)Graph` making use of
 
 For a more comprehensive exploration of the package features and functionalities we strongly recommend consulting the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs.jl).  
 
+# Installation and Usage 
+
+To install MultilayerGraphs.jl it is sufficient to activate the `pkg` mode by pressing `]` in the Julia REPL and then run the following command:
+
+```nothing
+pkg> add MultilayerGraphs
+```
+
+In the following code chunks we synthetically illustrate some of the main features outlined in the previous section. 
+
+Let's begin by importing the necessary dependencies and setting the relevant constants.
+
+```julia
+# Import necessary dependencies
+using Distributions, Graphs, SimpleValueGraphs
+using MultilayerGraphs
+# Set the number of nodes
+const n_nodes = 100 
+# Create a list of nodes
+const node_list = [Node("node_$i") for i in 1:n_nodes]
+```
+
+## Layers and Interlayers 
+
+We will instantiate layers and interlayers with randomly-selected edges and vertices adopting a variety of techniques. Layers and Interlayers are not immutable, and mostly behave like normal graphs. The reader is invited to consult the [API](https://juliagraphs.org/MultilayerGraphs.jl/stable/API/) for more information.
+
+Here we define a layer with an underlying simple directed graph using a graph generator-like (or "configuration model"-like) constructor which allows us to specify both the **indegree** and the **outdegree sequences**. Before instantiating each layer we sample the number of its vertices and, optionally, of its edges.
+
+```julia
+# Create a simple directed layer
+n_vertices = rand(1:100)                          # Number of vertices 
+layer_simple_directed = layer_simpledigraph(      # Layer constructor 
+    :layer_simple_directed,                       # Layer name
+    sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
+    Truncated(Normal(5, 5), 0, 20), # Indegree sequence distribution 
+    Truncated(Normal(5, 5), 0, 20)  # Outdegree sequence distribution
+)
+```
+
+Then we define a layer with an underlying simple weighted directed graph. This is another kind of constructor that allows the user to specify the number of edges to be randomly distributed among vertices. 
+
+```julia
+# Create a simple directed weighted layer
+n_vertices = rand(1:n_nodes)                                   # Number of vertices 
+n_edges = rand(n_vertices:(n_vertices * (n_vertices - 1) - 1)) # Number of edges 
+layer_simple_directed_weighted = layer_simpleweighteddigraph(  # Layer constructor 
+    :layer_simple_directed_weighted,                           # Layer name
+    sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
+    n_edges;                                 # Number of randomly distributed edges
+    default_edge_weight=(src, dst) -> rand() # Function assigning weights to edges 
+)
+```
+
+Similar constructors, more flexible at the cost of ease of use, enable a finer tuning. The constructor we use below should be necessary only in rare circumstances, e.g. if the equivalent simplified constructor `layer_simplevaldigraph` is not able to infer the correct return types of `default_vertex_metadata` or `default_edge_metadata`, or to use and underlying graph structure that isn't currently supported.
+
+```julia
+# Create a simple directed value layer
+n_vertices = rand(1:n_nodes)                                   # Number of vertices 
+n_edges = rand(n_vertices:(n_vertices * (n_vertices - 1) - 1)) # Number of edges 
+default_vertex_metadata = v -> ("vertex_$(v)_metadata")        # Vertex metadata 
[…]
