### 2023-01-08 #5055 @pitmonticone
Dear editors, thank you very much for your availability and ready response with helpful feedbacks and clear instructions.

Following @danielskatz advice we (@ClaudMor and I) have decided to write three versions of the [paper](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper) with different lengths:
- [long version](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper_long.md) (2286 words): contains the "Installation and Usage" section as in the one previously submitted;
- [medium version](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper_medium.md) (1548 words): contains the "Installation and Usage" section with just a brief introduction before a single code chunk and less comments within it;
- [short version](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper_short.md) (1155 words): doesn't contain the "Installation and Usage" section and therefore doesn't include any code chunk.

We would be glad to hear from you which one we should ultimately submit.

Here are some potential reviewers extracted from your list:
- juliohm: "Julia" among preferred languages and "graph theory" among preferred domains;
- tpoisot: "Julia" among preferred languages and "graph theory" among preferred domains;
- Datseris: "Julia" among preferred languages and "complex systems" among preferred domains;
- bgailleton: "Julia" among other languages and "graph theory" among preferred domains;
- TomKellyGenetics: "Julia" among other languages and "graph theory" among preferred domains;
- amritagos: "Julia" among other languages and "graph theory" among preferred domains.

Thanks in advance.
