+++ src/synchronizedmultiplexgraph.jl
@@ -1,46 +1,51 @@
 """
-    MultilayerGraph{T, U, G <: AbstractGraph{T}} <: AbstractMultilayerGraph{T,U}
+    SynchronizedMultiplexGraph{T, U, G <: AbstractGraph{T}} <: AbstractMultilayerGraph{T,U}
 
-A concrete type that can represent a general multilayer graph. Its internal fields aren't meant to be modified by the user. Please prefer the provided API.
+A concrete type that can represent an undirected multiplex graph with the following special functionalities: 
+
+1. When adding a `Node` via `add_node!, a corresponding `MultilayerVertex` is created in all layers. `add_vertex!` and `remove_vertex!` are thus disabled.
+
+
+
+
+Its internal fields aren't meant to be modified by the user. Please prefer the provided API.
 """
-mutable struct MultilayerGraph{T,U} <: AbstractMultilayerUGraph{T,U}
+mutable struct SynchronizedMultiplexGraph{T,U} <: AbstractMultiplexUGraph{T,U}
     layers::Vector{LayerDescriptor{T,U}} # vector containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.
-    interlayers::OrderedDict{Set{Symbol},InterlayerDescriptor{T,U}} # the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
+    interlayers::OrderedDict{Set{Symbol},InterlayerDescriptor{T,U, SimpleGraph{T}}} # the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
     v_V_associations::Bijection{T,<:MultilayerVertex} # A Bijection from Bijections.jl that associates numeric vertices to `MultilayerVertex`s.
     idx_N_associations::Bijection{Int64,Node} # A Bijection from Bijections.jl that associates Int64 to `Node`s.
-    fadjlist::Vector{Vector{HalfEdge{<:MultilayerVertex,<:Union{Nothing,U}}}} # the forward adjacency list of the MultilayerGraph. It is a vector of vectors of `HalfEdge`s. Its i-th element are the `HalfEdge`s that originate from `v_V_associations[i]`.
+    fadjlist::Vector{Vector{HalfEdge{<:MultilayerVertex,<:Union{Nothing,U}}}} # the forward adjacency list of the SynchronizedMultiplexGraph. It is a vector of vectors of `HalfEdge`s. Its i-th element are the `HalfEdge`s that originate from `v_V_associations[i]`.
     v_metadata_dict::Dict{T,<:Union{<:Tuple,<:NamedTuple}} #  A Dictionary that associates numeric vertices to their metadata.
 end
 
 # Traits
-@traitimpl IsWeighted{MultilayerGraph}
-@traitimpl IsMeta{MultilayerGraph}
+@traitimpl IsWeighted{SynchronizedMultiplexGraph}
+@traitimpl IsMeta{SynchronizedMultiplexGraph}
 
 # Constructors
 """
-    MultilayerGraph(
-        layers::Vector{<:Layer{T,U}},
-        specified_interlayers::Vector{<:Interlayer{T,U}};
-        default_interlayers_null_graph::H = SimpleGraph{T}(),
-        default_interlayers_structure::String="multiplex",
+    SynchronizedMultiplexGraph(
+        layers::Vector{<:Layer{T,U}};
+        default_interlayers_null_graph::H = SimpleGraph{T}()
     ) where {T,U, H <: AbstractGraph{T}}   
 
[…]
