+++ src/abstractmultiplexgraph.jl
@@ -1,0 +1,6 @@
+"""
+    AbstractMultiplexGraph{T,U} <: AbstractMultilayerGraph{T,U}
+
+An abstract type representing a general multiplex graph.
+"""
+abstract type AbstractMultiplexGraph{T,U} <: AbstractMultilayerGraph{T,U} end

[…]
