### 2023-01-09 1d28cc5
Implement Multiplex(Di)Graph
TODO:

write tests

Co-Authored-By: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-Authored-By: Claudio Moroni <43729990+ClaudMor@users.noreply.github.com>
