+++ docs/make.jl
@@ -1,3 +1,4 @@
+using Revise
 using Distributions
 using Graphs
 using MultilayerGraphs
