### 2023-01-10 #5055 @diehlpk
Hi @pitmonticone

[medium version](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper_medium.md) (1548 words): contains the "Installation and Usage" section with just a brief introduction before a single code chunk and less comments within it;

I think that maybe the medium version could be the way to go. However, the code listing is very long and might be better in the example folder and just be linked from the paper. Same for the installation. Might I suggest that some graphical illustration to define the keywords?
