+++ example/example.jl
@@ -6,7 +6,7 @@
 using Distributions, Graphs, SimpleValueGraphs
 using MultilayerGraphs
 # Set the number of nodes
-const n_nodes = 100 
+const n_nodes = 100
 # Create a list of nodes
 const node_list = [Node("node_$i") for i in 1:n_nodes]
 
@@ -20,7 +20,7 @@ layer_simple_directed = layer_simpledigraph(      # Layer constructor
     :layer_simple_directed,                       # Layer name
     sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
     Truncated(Normal(5, 5), 0, 20), # Indegree sequence distribution 
-    Truncated(Normal(5, 5), 0, 20)  # Outdegree sequence distribution
+    Truncated(Normal(5, 5), 0, 20),  # Outdegree sequence distribution
 )
 
 # Create a simple directed weighted layer
@@ -30,7 +30,7 @@ layer_simple_directed_weighted = layer_simpleweighteddigraph(  # Layer construct
     :layer_simple_directed_weighted,                           # Layer name
     sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
     n_edges;                                 # Number of randomly distributed edges
-    default_edge_weight=(src, dst) -> rand() # Function assigning weights to edges 
+    default_edge_weight=(src, dst) -> rand(), # Function assigning weights to edges 
 )
 
 # Create a simple directed value layer
@@ -42,8 +42,8 @@ layer_simple_directed_value = Layer(                           # Layer construct
     :layer_simple_directed_value,                              # Layer name
     sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
     n_edges,                                      # Number of randomly distributed edges
-    ValDiGraph(                                                
-        SimpleDiGraph{Int64}(); 
+    ValDiGraph(
+        SimpleDiGraph{Int64}();
         vertexval_types=(String,),
         vertexval_init=default_vertex_metadata,
         edgeval_types=(Float64,),
@@ -51,11 +51,13 @@ layer_simple_directed_value = Layer(                           # Layer construct
     ),
     Float64;
     default_vertex_metadata=default_vertex_metadata, # Vertex metadata 
-    default_edge_metadata=default_edge_metadata      # Edge metadata 
+    default_edge_metadata=default_edge_metadata,      # Edge metadata 
 )
 
 # Create a list of layers 
-layers = [layer_simple_directed, layer_simple_directed_weighted, layer_simple_directed_value]
+layers = [
+    layer_simple_directed, layer_simple_directed_weighted, layer_simple_directed_value
+]
 
 #################################################
 ################# INTERLAYERS ###################
@@ -68,7 +70,7 @@ n_edges = rand(1:(n_vertices_1 * n_vertices_2 - 1))    # Number of interlayer ed
 interlayer_simple_directed = interlayer_simpledigraph( # Interlayer constructor 
     layer_simple_directed,                             # Layer 1 
     layer_simple_directed_weighted,                    # Layer 2 
-    n_edges                                            # Number of edges 
+    n_edges,                                            # Number of edges 
 )
 
 # Create a simple directed meta interlayer 
@@ -80,8 +82,8 @@ interlayer_simple_directed_meta = interlayer_metadigraph( # Interlayer construct
     layer_simple_directed_value,                          # Layer 2
     n_edges;                                              # Number of edges
     default_edge_metadata=(src, dst) ->                   # Edge metadata 
-        (edge_metadata="metadata_of_edge_from_$(src)_to_$(dst)"),
[…]
