+++ paper/paper.bib
@@ -519,3 +519,14 @@
   author = {Mark E. Dickison and Matteo Magnani and Luca Rossi},
   title = {Multilayer Social Networks}
 }
+
+# FIGURES 
+@article{DeDomenicoIllustrated2022,
+  doi = {10.17605/OSF.IO/GY53K},
+  url = {https://doi.org/10.17605/osf.io/gy53k},
+  author = {De Domenico,  Manlio},
+  title = {Multilayer Networks Illustrated},
+  publisher = {Open Science Framework},
+  year = {2022},
+  copyright = {Other}
+}
