+++ src/ismultiplex.jl
@@ -3,8 +3,8 @@
 
 
 # Layers and Interlayers
-@traitfn function specify_interlayer!(
+@traits function specify_interlayer!(
     mg::M, new_interlayer::In
-) where {T,U,G<:AbstractGraph{T},M<:AbstractMultilayerGraph{T,U},In<:Interlayer{T,U,G}; IsMultiplex{M}}
+) where {T,U,G<:AbstractGraph{T},M<:AbstractMultilayerGraph{T,U},In<:Interlayer{T,U,G}, istrait(IsMultiplex{M})}
      throw(ErrorException("Interlayers are automatically specified in multiplex graphs and they should not be modified in any way. If you need to modify interlayers, perhaps a standard multilayer graph will better suit your use case."))
 end
