+++ src/graphs_extensions/graphs_extensions.jl
@@ -5,13 +5,13 @@ include("metagraphs.jl")
 include("simplevaluegraphs.jl")
 
 """
-    _edges(g::Union{Graphs.SimpleGraphs.AbstractSimpleGraph{T}, AbstractSimpleWeightedGraph{T}, AbstractMetaGraph{T}, SimpleValueGraphs.AbstractValGraph{T} }, weighttype::Type{U}) where {T,U}
+    _edges(g::Union{SimpleGraphs.AbstractSimpleGraph{T}, AbstractSimpleWeightedGraph{T}, AbstractMetaGraph{T}, SimpleValueGraphs.AbstractValGraph{T} }, weighttype::Type{U}) where {T,U}
 
 Internal function. It serves as a unified interface between MultilayerGraphs.jl `edges` method and the homonymous methods of the other packages.
 """
 function _edges(
     g::Union{
-        Graphs.SimpleGraphs.AbstractSimpleGraph{T},
+        SimpleGraphs.AbstractSimpleGraph{T},
         AbstractSimpleWeightedGraph{T},
         AbstractMetaGraph{T},
         SimpleValueGraphs.AbstractValGraph{T},
