+++ src/abstractmultilayergraph.jl
@@ -10,7 +10,7 @@
 
 # Base.eltype
 # weighttype
-# Graphs.edgetype
+# edgetype
 
 # has_vertex
 # multilayer_vertices
@@ -132,7 +132,7 @@ Base.eltype(::M) where {T,M<:AbstractMultilayerGraph{T}} = T
 
 Return true if `v` is in mg, else false.
 """
-function Graphs.has_vertex(mg::M, v::T) where {T,M<:AbstractMultilayerGraph{T}}
+function has_vertex(mg::M, v::T) where {T,M<:AbstractMultilayerGraph{T}}
     return v in domain(mg.v_V_associations)
 end
 
@@ -141,7 +141,7 @@ end
 
 Return true if `mv` is in `mg`, else false.
 """
-function Graphs.has_vertex(mg::AbstractMultilayerGraph, mv::MultilayerVertex)
+function has_vertex(mg::AbstractMultilayerGraph, mv::MultilayerVertex)
     return get_bare_mv(mv) in image(mg.v_V_associations)
 end
 
@@ -157,14 +157,14 @@ mv_vertices(mg::AbstractMultilayerGraph) = [get_rich_mv(mg, v) for v in vertices
 
 Return the number of vertices in `mg`, excluding the missing vertices.
 """
-Graphs.nv(mg::AbstractMultilayerGraph) = length(mg.v_V_associations)
+nv(mg::AbstractMultilayerGraph) = length(mg.v_V_associations)
 
 """
     vertices(mg::M) where {M<:AbstractMultilayerGraph}
 
 Return the collection of the vertices of `mg`.
 """
-function Graphs.vertices(mg::M) where {M<:AbstractMultilayerGraph}
+function vertices(mg::M) where {M<:AbstractMultilayerGraph}
     return sort(collect(domain(mg.v_V_associations)))
 end
 
@@ -198,7 +198,7 @@ end
 
 Add MultilayerVertex `mv` to multilayer graph `mg`. If `add_node` is true and `node(mv)` is not already part of `mg`, then add `node(mv)` to `mg` before adding `mv` to `mg` instead of throwing an error.
 """
-function Graphs.add_vertex!(
+function add_vertex!(
     mg::AbstractMultilayerGraph, mv::MultilayerVertex; add_node::Bool=true
 )
     _node = node(mv)
@@ -214,21 +214,21 @@ end
 
 Return the edge type for `mg`.
 """
-Graphs.edgetype(::M) where {T,U,M<:AbstractMultilayerGraph{T,U}} = MultilayerEdge{U}
+edgetype(::M) where {T,U,M<:AbstractMultilayerGraph{T,U}} = MultilayerEdge{U}
 
 """
     ne(mg::AbstractMultilayerGraph)
 
 Return the number of edges in `mg`.
 """
-Graphs.ne(mg::AbstractMultilayerGraph) = length(edges(mg))
+ne(mg::AbstractMultilayerGraph) = length(edges(mg))
 
 """
     has_edge(mg::AbstractMultilayerGraph, edge::MultilayerEdge) 
 
 Return true if `mg` has an edge between the source and the destination of `edge` (does not check edge or vertex metadata).
 """
-function Graphs.has_edge(mg::AbstractMultilayerGraph, edge::MultilayerEdge)
+function has_edge(mg::AbstractMultilayerGraph, edge::MultilayerEdge)
     return has_edge(mg, get_v(mg, src(edge)), get_v(mg, dst(edge)))
 end
 
@@ -237,7 +237,7 @@ end
 
 Return true if `mg` has edge between the `src` and `dst` (does not check edge or vertex metadata).
 """
-function Graphs.has_edge(
[…]
