### 2023-01-10 #5055 @pitmonticone
Hi @diehlpk,

Thank you very much for your feedback.

> I think that maybe the medium version could be the way to go. However, the code listing is very long and might be better in the example folder and just be linked from the paper. Same for the installation. Might I suggest that some graphical illustration to define the keywords?

We (@ClaudMor and I) have just updated the [paper](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/JOSS/paper/paper.md) by:
- linking the files in the [example folder](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/main/example);
- adding two [images](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/JOSS/paper/images): one for the package logo and one for a diagram depicting a multi-scale biological application.
