+++ proposal/proposal.md
@@ -1,0 +1,87 @@
+# JuliaCon 2023 Proposal 
+
+## Title 
+
+MultilayerGraphs.jl: Multilayer Network Science in Julia
+
+## Session Type 
+
+Lightning talk (10 minutes)
+
+## Track 
+
+JuliaCon 
+
+## Abstract 
+
+**MultilayerGraphs.jl** is a Julia package for the creation, manipulation and analysis of multilayer graphs, which have been adopted to model of a wide range of complex systems from bio-chemical to socio-technical networks.
+
+## Description 
+
+[**MultilayerGraphs.jl**](https://github.com/JuliaGraphs/MultilayerGraphs.jl) is a Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs. 
+
+A multilayer graph consists of multiple subgraphs called **layers** which can be interconnected through [bipartite graphs](https://en.wikipedia.org/wiki/Bipartite_graph) called **interlayers**. 
+
+Several theoretical frameworks have been proposed to formally subsume all instances of multilayer graphs ([De Domenico  et al. (2013)](https://doi.org/10.1103/physrevx.3.041022); [Kivelä et al. (2014)](https://doi.org/10.1093/comnet/cnu016); [Bianconi (2018)](https://doi.org/10.1093/oso/9780198753919.001.0001); [De Domenico (2022)](https://doi.org/10.1007/978-3-030-75718-2)). 
+
+Multilayer graphs have been adopted to model the structure and dynamics of a wide spectrum of high-dimensional, non-linear, multi-scale, time-dependent complex systems including physical, chemical, biological, neuronal, socio-technical, epidemiological, ecological and economic networks ([Cozzo et al. (2013)](https://doi.org/10.1103/physreve.88.050801); [Estrada and Gomez-Gardenes (2014)](https://doi.org/10.1103/physreve.89.042819); [Baggio et al. (2016)](https://doi.org/10.1073/pnas.1604401113); [DeDomenico et al. (2016)](https://doi.org/10.1038/nphys3865); [Amato et al. (2017)](https://doi.org/10.1038/s41598-017-06933-2); [DeDomenico (2017)](https://doi.org/10.1093/gigascience/gix004); [Timteo et al. (2018)](https://doi.org/10.1038/s41467-017-02658-y); [Aleta et al. (2020)](https://doi.org/10.1038/s41562-020-0931-9); [Aleta et al. (2022)](https://doi.org/10.1073/pnas.2112182119)). 
+
+MultilayerGraphs.jl is integrated within the **JuliaGraphs** ecosystem extending [Graphs.jl](https://github.com/JuliaGraphs/Graphs.jl) with several multilayer-specific methods and metrics including the multilayer eigenvector centrality, the multilayer modularity and the Von Newman entropy. 
+
+The package features an implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects a practitioner would expect such as nodes (`Node`), vertices (`MultilayerVertex`), layers (`Layer`), interlayers (`Interlayer`), etc.
+
+The two main data structures are collections of layers connected through interlayers called `MultilayerGraph` and `MultilayerDiGraph`.
+
+The **vertices** of a multilayer graph are representations of one set of distinct objects called `Node`s. Each layer may represent all the node set or just a subset of it. The vertices of `Multilayer(Di)Graph` are implemented via the `MultilayerVertex` custom type. Each `MultilayerVertex` encodes information about the node it represents, the layer it belongs to and its metadata. 
+
[…]
