+++ proposal/README.md
@@ -1,0 +1,58 @@
+# WAW 2023: 18th Workshop on Algorithms and Models for the Web Graph
+
+## Where 
+
+[Fields Institute for Research in Mathematical Sciences](http://www.fields.utoronto.ca), Toronto, Canada. 
+
+## When 
+
+2023/05/23 - 2023/05/26
+
+## What 
+
+The World Wide Web has become part of our everyday life, and information retrieval and data mining on the Web are now of enormous practical interest. The algorithms supporting these activities combine the view of the Web as a text repository and as a graph, induced in various ways by links among pages, hosts and users.
+
+The aim of the 18th Workshop on Algorithms and Models for the Web Graph (WAW 2023) is to further the understanding of graphs that arise from the Web and various user activities on the Web, and stimulate the development of high-performance algorithms and applications that exploit these graphs. The workshop will also welcome the researchers who are working on graph-theoretic and algorithmic aspects of citation networks, social networks, biological networks, molecular networks, and the Internet.
+
+WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks. 
+
+More details can be found [here](https://math.ryerson.ca/waw2023/).
+
+### Contents 
+
+- **Algorithms**: graph algorithms, clustering, collaborative filtering, routing optimization.
+- **Analysis**: structural properties, visualization, patterns, communities, discovery, flow simulation.
+- **Data Models**: hypergraph models, graph models, evolution, trust and reputation networks.
+- **Topics**: Web, social networks, transportation networks, communication networks, recommender networks, citation networks, Wikipedia, biological networks, blogs, p2p.
+- **Applications**: web mining, social applications, routing & transportation, web search and ranking.
+
+## Submission Guidelines 
+
+### Papers
+
+WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks.
+
+The papers must be formatted according to Springer LNCS style. The maximum length of papers is at most 15 pages but shorter papers are welcome. The workshop proceedings will be published as a Springer LNCS volume. Selected papers from the workshop will be invited to a special issue. Talks associated with proceedings papers will be 30 minutes.
+
+### Abstracts 
+
+Abstracts are a maximum of one page, may be in any style, and must be pdf files. They should include the title, speakers, and a summary of research. All abstracts related to the conference areas will be accepted (at the discretion of the program committee), and published on the conference website (but not published in the proceedings). Talks associated with abstracts will be 20 minutes.
+
+Papers and abstracts must be submitted via Easychair.
+
+## Publication 
+
+WAW 2023 proceedings will be published by Springer LNCS volume. 
+
+## Contact 
+
+All questions about submissions should be emailed to Pawel Pralat (pralat@ryerson.ca).
+
+## Sponsors 
+
+- Fields Institute 
+- NAWA
+- SGH Warsaw School of Economics
+- Toronto Metropolitan University
+- Tutte Institute
+- Google
