+++ proposal/README.md
@@ -19,11 +19,9 @@
 
 The World Wide Web has become part of our everyday life, and information retrieval and data mining on the Web are now of enormous practical interest. The algorithms supporting these activities combine the view of the Web as a text repository and as a graph, induced in various ways by links among pages, hosts and users.
 
-The aim of the 18th Workshop on Algorithms and Models for the Web Graph (WAW 2023) is to further the understanding of graphs that arise from the Web and various user activities on the Web, and stimulate the development of high-performance algorithms and applications that exploit these graphs. The workshop will also welcome the researchers who are working on graph-theoretic and algorithmic aspects of citation networks, social networks, biological networks, molecular networks, and the Internet.
+### Aim
 
-WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks. 
-
-More details can be found [here](https://math.ryerson.ca/waw2023/).
+To further the understanding of graphs that arise from the Web and various user activities on the Web, and stimulate the development of high-performance algorithms and applications that exploit these graphs. The workshop will also welcome the researchers who are working on graph-theoretic and algorithmic aspects of citation networks, social networks, biological networks, molecular networks, and the Internet.
 
 ### Contents 
 
@@ -33,21 +31,21 @@ More details can be found [here](https://math.ryerson.ca/waw2023/).
 - **Topics**: Web, social networks, transportation networks, communication networks, recommender networks, citation networks, Wikipedia, biological networks, blogs, p2p.
 - **Applications**: web mining, social applications, routing & transportation, web search and ranking.
 
-## Submission Guidelines 
+### Submission Guidelines 
 
-### Papers
+#### Papers
 
 WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks.
 
 The papers must be formatted according to Springer LNCS style. The maximum length of papers is at most 15 pages but shorter papers are welcome. The workshop proceedings will be published as a Springer LNCS volume. Selected papers from the workshop will be invited to a special issue. Talks associated with proceedings papers will be 30 minutes.
 
-### Abstracts 
+#### Abstracts 
 
 Abstracts are a maximum of one page, may be in any style, and must be pdf files. They should include the title, speakers, and a summary of research. All abstracts related to the conference areas will be accepted (at the discretion of the program committee), and published on the conference website (but not published in the proceedings). Talks associated with abstracts will be 20 minutes.
 
 Papers and abstracts must be submitted via Easychair.
 
-## Publication 
+### Publication 
 
 WAW 2023 proceedings will be published by Springer LNCS volume. 
 
