+++ proposal/README.md
@@ -1,5 +1,7 @@
 # WAW 2023: 18th Workshop on Algorithms and Models for the Web Graph
 
+![poster](https://math.ryerson.ca/waw2023/images/WAW2023-poster.png)
+
 ## Main References 
 
 - [Conference Website](https://math.ryerson.ca/waw2023/);
@@ -23,6 +25,8 @@ The World Wide Web has become part of our everyday life, and information retriev
 
 To further the understanding of graphs that arise from the Web and various user activities on the Web, and stimulate the development of high-performance algorithms and applications that exploit these graphs. The workshop will also welcome the researchers who are working on graph-theoretic and algorithmic aspects of citation networks, social networks, biological networks, molecular networks, and the Internet.
 
+WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks.
+
 ### Contents 
 
 - **Algorithms**: graph algorithms, clustering, collaborative filtering, routing optimization.
@@ -33,17 +37,15 @@ To further the understanding of graphs that arise from the Web and various user 
 
 ### Submission Guidelines 
 
-#### Papers
+Papers and abstracts must be submitted via [Easychair](https://easychair.org/conferences/?conf=waw2023).
 
-WAW 2023 invites original research papers and abstracts on all aspects of algorithmic and mathematical research in the areas pertaining to the World-Wide Web, especially papers that espouse the view of complex data as networks.
+#### Papers
 
 The papers must be formatted according to Springer LNCS style. The maximum length of papers is at most 15 pages but shorter papers are welcome. The workshop proceedings will be published as a Springer LNCS volume. Selected papers from the workshop will be invited to a special issue. Talks associated with proceedings papers will be 30 minutes.
 
 #### Abstracts 
 
 Abstracts are a maximum of one page, may be in any style, and must be pdf files. They should include the title, speakers, and a summary of research. All abstracts related to the conference areas will be accepted (at the discretion of the program committee), and published on the conference website (but not published in the proceedings). Talks associated with abstracts will be 20 minutes.
-
-Papers and abstracts must be submitted via Easychair.
 
 ### Publication 
 
