+++ proposal/README.md
@@ -1,6 +1,6 @@
 # WAW 2023: 18th Workshop on Algorithms and Models for the Web Graph
 
-![poster](https://math.ryerson.ca/waw2023/images/WAW2023-poster.png)
+![poster](https://math.ryerson.ca/waw2023/images/WAW2023-poster-small.png)
 
 ## Main References 
 
