+++ proposal/abstract.md
@@ -1,0 +1,22 @@
+# Abstract
+
+## Title 
+
+MultilayerGraphs.jl: Multilayer Network Science in Julia
+
+## Speakers 
+
+- Claudio Moroni (University of Turin);
+- Pietro Monticone (University of Turin).
+
+## Text 
+
+[**MultilayerGraphs.jl**](https://github.com/JuliaGraphs/MultilayerGraphs.jl) is a Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs. 
+
+A multilayer graph consists of multiple subgraphs called **layers** which can be interconnected through bipartite graphs called **interlayers**. 
+
+In order to formally represent multilayer networks, multiple theoretical frameworks have been proposed and adopted to model a wide spectrum of high-dimensional, non-linear, multi-scale, time-dependent complex systems including molecular, neuronal, socio-technical and economic networks. 
+
+MultilayerGraphs.jl is integrated within the [**JuliaGraphs**](https://github.com/JuliaGraphs) ecosystem extending [Graphs.jl](https://github.com/JuliaGraphs/Graphs.jl) with several multilayer-specific methods and metrics and within the [**JuliaDynamics**](https://github.com/JuliaDynamics) ecosystem allowing for agent-based modelling on general multilayer networks via [Agents.jl](https://github.com/JuliaDynamics/Agents.jl).
+
+For a comprehensive exploration of its features and functionalities we recommend to consult the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs).
