+++ proposal/bibliography.bib
@@ -1,0 +1,532 @@
+# PACKAGES 
+
+## Julia 
+
+@article{Bezanson2017,
+  doi = {10.1137/141000671},
+  url = {https://doi.org/10.1137/141000671},
+  year = {2017},
+  month = jan,
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume = {59},
+  number = {1},
+  pages = {65--98},
+  author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title = {Julia: A Fresh Approach to Numerical Computing},
+  journal = {{SIAM} Review}
+}
+
+@misc{Graphs2021,
+  author = {Fairbanks, James and Besan{\c{c}}on, Mathieu and Simon, Sch{\"o}lly and Hoffiman, J{\'u}lio and Eubank, Nick and Karpinski, Stefan},
+  title = {JuliaGraphs/Graphs.jl: an optimized graphs package for the Julia programming language},
+  year = 2021,
+  url = {https://github.com/JuliaGraphs/Graphs.jl/}
+}
+
+@article{Datseris2022,
+  doi = {10.1177/00375497211068820},
+  url = {https://doi.org/10.1177/00375497211068820},
+  year = {2022},
+  month = jan,
+  publisher = {{SAGE} Publications},
+  pages = {003754972110688},
+  author = {George Datseris and Ali R. Vahdati and Timothy C. DuBois},
+  title = {Agents.jl: a performant and feature-full agent-based modeling software of minimal code complexity},
+  journal = {{SIMULATION}}
+}
+
+@software{Moroni_Monticone_MultilayerGraphs_2022,
+         abstract     = {A Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs.},
+         author       = {Moroni, Claudio and Monticone, Pietro},
+         doi          = {10.5281/zenodo.7009172},
+         institution  = {University of Turin (UniTO)},
+         keywords     = {Julia Language, Julia Package, Graph Theory, Applied Graph Theory, Network Theory, Network Science, Graphs, Multilayer Graphs, Multilevel Graphs, Hierarchical Graphs, Networks, Multilayer Networks, Multilevel Networks, Hierarchical Networks, Complexity, Complex Systems},
+         license      = {MIT},
+         organization = {Interdisciplinary Physics Team (InPhyT)},
+         title        = {MultilayerGraphs.jl: A Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs},
+         url          = {https://doi.org/10.5281/zenodo.7009172},
+         year         = {2022}
+         }
+
+## Python 
+
+## R 
+@article{Lazega2008,
+  doi = {10.1016/j.socnet.2008.02.001},
+  url = {https://doi.org/10.1016/j.socnet.2008.02.001},
+  year = {2008},
+  month = may,
+  publisher = {Elsevier {BV}},
+  volume = {30},
+  number = {2},
+  pages = {159--176},
+  author = {Emmanuel Lazega and Marie-Th{\'{e}}r{\`{e}}se Jourda and Lise Mounier and Rafaël Stofer},
+  title = {Catching up with big fish in the big pond? Multi-level network analysis through linked design},
+  journal = {Social Networks}
+}
+
+@article{DeDomenico2014,
+  doi = {10.1093/comnet/cnu038},
+  url = {https://doi.org/10.1093/comnet/cnu038},
+  year = {2014},
+  month = oct,
[…]
