+++ proposal/paper/README.md
@@ -1,0 +1,2 @@
+# Paper 
+

[…]
