+++ proposal/abstract/abstract.md
@@ -11,12 +11,17 @@ Multilayer Network Science in Julia with MultilayerGraphs.jl
 
 ## Text 
 
-[**MultilayerGraphs.jl**](https://github.com/JuliaGraphs/MultilayerGraphs.jl) is a Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs. 
+[**MultilayerGraphs.jl**](https://github.com/JuliaGraphs/MultilayerGraphs.jl) is a Julia package for the creation, manipulation and analysis of the structure, dynamics and functions of multilayer graphs.
 
-A multilayer graph consists of multiple subgraphs called **layers** which can be interconnected through bipartite graphs called **interlayers**. 
+A multilayer graph consists of multiple subgraphs called **layers** which can be interconnected through bipartite graphs called **interlayers** composed of the sets of vertices of two different layers and the edges between them.
 
-In order to formally represent multilayer networks, multiple theoretical frameworks have been proposed and adopted to model a wide spectrum of high-dimensional, non-linear, multi-scale, time-dependent complex systems including molecular, neuronal, socio-technical and economic networks. 
+In order to formally represent multilayer networks, multiple theoretical paradigms 
+have been proposed and adopted to model a wide spectrum of high-dimensional, 
+multi-scale, time-dependent complex systems including molecular,
+neuronal, social, ecological and economic networks.
+
+The package features an implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects a practitioner would expect such as nodes, vertices, layers, interlayers, etc.
 
 MultilayerGraphs.jl is integrated within the [**JuliaGraphs**](https://github.com/JuliaGraphs) ecosystem extending [Graphs.jl](https://github.com/JuliaGraphs/Graphs.jl) with several multilayer-specific methods and metrics and within the [**JuliaDynamics**](https://github.com/JuliaDynamics) ecosystem allowing for agent-based modelling on general multilayer networks via [Agents.jl](https://github.com/JuliaDynamics/Agents.jl).
 
-For a comprehensive exploration of its features and functionalities we recommend to consult the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs).
+For a comprehensive exploration of the package features and functionalities the readers is invited to consult the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs).
