+++ proposal/abstract/abstract.tex
@@ -40,24 +40,40 @@ University of Turin, Italy (\url{https://www.unito.it}) \\ \and
 Interdisciplinary Physics Team, Italy (\url{https://github.com/InPhyT})
 }
 %
-\maketitle              % typeset the header of the contribution
+\maketitle % typeset the header of the contribution
 %
+% The abstract should briefly summarize the contents of the paper in
+% 150--250 words.
 \begin{abstract}
-The abstract should briefly summarize the contents of the paper in
-150--250 words.
+    
+MultilayerGraphs.jl is a Julia package for the creation, manipulation and analysis 
+of the structure, dynamics and functions of multilayer graphs. \\
 
-MultilayerGraphs.jl is a Julia package for the creation, manipulation 
-and analysis of the structure, dynamics and functions of multilayer graphs. 
 A multilayer graph consists of multiple subgraphs called layers which 
-can be interconnected through bipartite graphs called interlayers. 
+can be interconnected through bipartite graphs called interlayers composed 
+of the sets of vertices of two different layers and the edges between them. \\
 
-In order to formally represent multilayer networks, multiple theoretical frameworks have been proposed and adopted to model a wide spectrum of high-dimensional, non-linear, multi-scale, time-dependent complex systems including molecular, neuronal, socio-technical and economic networks. 
+In order to formally represent multilayer networks, multiple theoretical paradigms 
+have been proposed and adopted to model a wide spectrum of high-dimensional, 
+multi-scale, time-dependent complex systems including molecular,
+neuronal, social, ecological and economic networks. \\
 
-The package features an implementation that maps a standard integer-labelled vertex representation to a more user-friendly framework exporting all the objects a practitioner would expect such as nodes, vertices, layers, interlayers, etc.
+The package features an implementation that maps a standard integer-labelled 
+vertex representation to a more user-friendly framework exporting all the objects 
+a practitioner would expect such as nodes, vertices, layers, interlayers, etc. \\
 
-MultilayerGraphs.jl is integrated within the [**JuliaGraphs**](https://github.com/JuliaGraphs) ecosystem extending [Graphs.jl](https://github.com/JuliaGraphs/Graphs.jl) with several multilayer-specific methods and metrics and within the [**JuliaDynamics**](https://github.com/JuliaDynamics) ecosystem allowing for agent-based modelling on general multilayer networks via [Agents.jl](https://github.com/JuliaDynamics/Agents.jl).
+MultilayerGraphs.jl has been integrated within the JuliaGraph\footnote{https://github.com/JuliaGraphs.} 
+and the JuliaDynamics\footnote{https://github.com/JuliaDynamics.} ecosystems through: 
+\begin{itemize}
+    \item the extension of Graphs.jl\footnote{https://github.com/JuliaGraphs/Graphs.jl.} with several methods and metrics including 
+    the multilayer eigenvector centrality, the multilayer modularity and the Von Newman entropy; 
+    \item the compatibility with Agents.jl\footnote{https://github.com/JuliaDynamics/Agents.jl.} allowing for agent-based modelling 
+    on general multilayer networks. \\
+\end{itemize}
 
-For a comprehensive exploration of its features and functionalities we recommend to consult the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs).
+For a comprehensive exploration of the package features and functionalities the reader 
+is invited to consult the README\footnote{https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md.} 
+and documentation\footnote{https://juliagraphs.org/MultilayerGraphs.jl.}.
 
 \keywords{
     Discrete Mathematics \and Graph Theory \and Network Science \and \\ 
