+++ proposal/paper/paper.tex
@@ -62,7 +62,7 @@ The package features an implementation that maps a standard integer-labelled
 vertex representation to a more user-friendly framework exporting all the objects 
 a practitioner would expect such as nodes, vertices, layers, interlayers, etc. \\
 
-MultilayerGraphs.jl has been integrated within the JuliaGraph\footnote{https://github.com/JuliaGraphs.} 
+MultilayerGraphs.jl has been integrated within the JuliaGraphs\footnote{https://github.com/JuliaGraphs.} 
 and the JuliaDynamics\footnote{https://github.com/JuliaDynamics.} ecosystems through: 
 \begin{itemize}
     \item the extension of Graphs.jl\footnote{https://github.com/JuliaGraphs/Graphs.jl.} with several methods and metrics including 
