+++ proposal/paper/paper.bib
@@ -3,7 +3,6 @@
 ## Julia 
 
 @article{Bezanson2017,
-  doi = {10.1137/141000671},
   url = {https://doi.org/10.1137/141000671},
   year = {2017},
   month = jan,
@@ -24,7 +23,6 @@
 }
 
 @article{Datseris2022,
-  doi = {10.1177/00375497211068820},
   url = {https://doi.org/10.1177/00375497211068820},
   year = {2022},
   month = jan,
@@ -52,7 +50,6 @@
 
 ## R 
 @article{Lazega2008,
-  doi = {10.1016/j.socnet.2008.02.001},
   url = {https://doi.org/10.1016/j.socnet.2008.02.001},
   year = {2008},
   month = may,
@@ -66,7 +63,6 @@
 }
 
 @article{DeDomenico2014,
-  doi = {10.1093/comnet/cnu038},
   url = {https://doi.org/10.1093/comnet/cnu038},
   year = {2014},
   month = oct,
@@ -80,7 +76,6 @@
 }
 
 @article{Hammoud2018,
-  doi = {10.3390/genes9110519},
   url = {https://doi.org/10.3390/genes9110519},
   year = {2018},
   month = oct,
@@ -94,7 +89,6 @@
 }
 
 @article{Magnani2021,
-  doi = {10.18637/jss.v098.i08},
   url = {https://doi.org/10.18637/jss.v098.i08},
   year = {2021},
   publisher = {Foundation for Open Access Statistic},
@@ -108,7 +102,6 @@
 # ARTICLES  
 ## General 
 @article{DeDomenico2013,
-  doi = {10.1103/physrevx.3.041022},
   url = {https://doi.org/10.1103/physrevx.3.041022},
   year = {2013},
   month = dec,
@@ -121,7 +114,6 @@
 }
 
 @article{Kivela2014,
-  doi = {10.1093/comnet/cnu016},
   url = {https://doi.org/10.1093/comnet/cnu016},
   year = {2014},
   month = jul,
@@ -135,7 +127,6 @@
 }
 
 @article{Boccaletti2014,
-  doi = {10.1016/j.physrep.2014.07.001},
   url = {https://doi.org/10.1016/j.physrep.2014.07.001},
   year = {2014},
   month = nov,
@@ -149,7 +140,6 @@
 }
 
 @article{Lee2015,
-  doi = {10.1140/epjb/e2015-50742-1},
   url = {https://doi.org/10.1140/epjb/e2015-50742-1},
   year = {2015},
   month = feb,
@@ -162,7 +152,6 @@
 }
 
 @article{Aleta2019,
-  doi = {10.1146/annurev-conmatphys-031218-013259},
   url = {https://doi.org/10.1146/annurev-conmatphys-031218-013259},
   year = {2019},
   month = mar,
@@ -178,7 +167,6 @@
 ## Algorithms 
 ### Graph Realisation 
 @article{Hakimi1962,
-  doi = {10.1137/0110037},
   url = {https://doi.org/10.1137/0110037},
   year = {1962},
   month = sep,
@@ -192,7 +180,6 @@
 }
 
 @article{Kleitman1973,
-  doi = {10.1016/0012-365x(73)90037-x},
   url = {https://doi.org/10.1016/0012-365x(73)90037-x},
   year = {1973},
   month = sep,
@@ -208,7 +195,6 @@
 ## Applications 
 
 @article{Granell2013,
-  doi = {10.1103/physrevlett.111.128701},
   url = {https://doi.org/10.1103/physrevlett.111.128701},
   year = {2013},
   month = sep,
@@ -221,7 +207,6 @@
 }
 
 @article{Cozzo2013,
-  doi = {10.1103/physreve.88.050801},
   url = {https://doi.org/10.1103/physreve.88.050801},
[…]
