+++ paper/paper.bib
@@ -468,7 +468,7 @@
   month = jul,
   publisher = {Oxford University Press},
   author = {Ginestra Bianconi},
-  title = {Multilayer Networks}
+  title = {Multilayer Networks: Structure and Function}
 }
 
 @book{Cozzo2018,
@@ -477,7 +477,7 @@
   year = {2018},
   publisher = {Springer International Publishing},
   author = {Emanuele Cozzo and Guilherme Ferraz de Arruda and Francisco Aparecido Rodrigues and Yamir Moreno},
-  title = {Multiplex Networks}
+  title = {Multiplex Networks: Basic Formalism and Structural Properties}
 }
 
 @book{Artime2022,
@@ -487,7 +487,7 @@
   month = aug,
   publisher = {Cambridge University Press},
   author = {Oriol Artime and Barbara Benigni and Giulia Bertagnolli and Valeria d{\textquotesingle}Andrea and Riccardo Gallotti and Arsham Ghavasieh and Sebastian Raimondo and Manlio De Domenico},
-  title = {Multilayer Network Science}
+  title = {Multilayer Network Science: From Cells to Societies}
 }
 
 @book{DeDomenico2022,
