### 2023-01-13 b3cdad0
Create abstract.txt
