+++ proposal/paper/paper.tex
@@ -153,11 +153,13 @@ convention are also acceptable.
 
 \cite{DeDomenico2013,Kivela2014,Bianconi2018}
 
-\subsubsection{Acknowledgements} Please place your acknowledgments at
-the end of the paper, preceded by an unnumbered run-in heading (i.e.
-3rd-level heading).
+\subsubsection{Acknowledgements} 
 
-%
+Write here something about JuliaLang, JuliaGraphs, etc.
+
+This open-source research software project received no financial support.
+
+
 % ---- Bibliography ----
 %
 % BibTeX users should specify bibliography style 'splncs04'.
