+++ src/graphs_extensions/simplevaluegraphs.jl
@@ -81,3 +81,6 @@ function _set_metadata!(
 end
 
 weights(g::SimpleValueGraphs.AbstractValGraph{T}) where {T} = adjacency_matrix(g)
+
+
+is_directed(g::G) where {G <: SimpleValueGraphs.AbstractValGraph} = is_directed(G)
