+++ proposal/presentation/Project.toml
@@ -1,0 +1,36 @@
+[deps]
+Base64 = "2a0f44e3-6c83-55bd-87e4-b1978d98bd5f"
+BinaryProvider = "b99e7846-7c00-51b0-8f62-c81ae34c0232"
+Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+DefaultApplication = "3f0dd361-4fe0-5fc6-8523-80b14ec94d85"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+DocumenterMarkdown = "997ab1e6-3595-5248-9280-8efb232c3433"
+Glob = "c27321d9-0574-5035-807b-f59d2c89b15c"
+Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
+IOCapture = "b5f81e59-6552-4d32-b1f0-c071b021bf89"
+InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+LibGit2 = "76f85450-5226-5b5a-8eaa-529ad045b433"
+Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
+Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
+Mmap = "a63ad114-7e13-5084-954f-fe012c677804"
+MultilayerGraphs = "af984530-d72f-4904-8b81-b992bbbeaf3f"
+NetworkOptions = "ca575930-c2e3-43a9-ace4-1e988b2c1908"
+Parsers = "69de0a69-1ddd-5017-9359-2bf0b02dc9f0"
+Preferences = "21216c6a-2e73-6563-6e65-726566657250"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+REPL = "3fa0cd96-eef1-5676-8a61-b3b8758bbffb"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Remark = "79b45036-8e38-5d04-8f49-b9fb23ff5a0d"
+SHA = "ea8e919c-243c-51af-8825-aaa63cd721ce"
+Serialization = "9e88b42a-f829-5b0c-bbe9-9e923198166b"
+SimpleValueGraphs = "b43c691f-cac2-5415-8122-396fe16a49fc"
+SnoopPrecompile = "66db9d55-30c0-4569-8b51-7e840670fc0c"
+Sockets = "6462fe0b-24de-5631-8697-dd941f90decc"
+TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
