+++ src/IsMultiplexUndirected.jl
@@ -1,0 +1,3 @@
+#= """
+"""
+abstract type AbstractMultiplexUGraph{T,U} <: AbstractMultilayerUGraph{T,U} end =#
