+++ src/abstractmultiplexdigraph.jl
@@ -1,3 +1,0 @@
-"""
-"""
-abstract type AbstractMultiplexDiGraph{T,U} <: AbstractMultilayerDiGraph{T,U} end
