+++ src/graphs_extensions/simplevaluegraphs.jl
@@ -83,4 +83,4 @@ end
 weights(g::SimpleValueGraphs.AbstractValGraph{T}) where {T} = adjacency_matrix(g)
 
 
-is_directed(g::G) where {G <: SimpleValueGraphs.AbstractValGraph} = is_directed(G)
+# is_directed(g::G) where {G <: SimpleValueGraphs.AbstractValGraph} = is_directed(G)
