+++ src/graphs_extensions/metagraphs.jl
@@ -86,4 +86,4 @@ end
 
 weights(g::AbstractMetaGraph{T}) where {T} = adjacency_matrix(g)
 
-is_directed(g::G) where {G <: AbstractMetaGraph} = is_directed(G)
+# is_directed(g::G) where {G <: AbstractMetaGraph} = is_directed(G)
