+++ src/abstractmultiplexugraph.jl
@@ -1,3 +1,0 @@
-"""
-"""
-abstract type AbstractMultiplexUGraph{T,U} <: AbstractMultilayerUGraph{T,U} end
