+++ src/IsMultiplexDirected.jl
@@ -1,0 +1,3 @@
+#= """
+"""
+abstract type AbstractMultiplexDiGraph{T,U} <: AbstractMultilayerDiGraph{T,U} end =#
