+++ src/abstractmultilayerugraph.jl
@@ -13,9 +13,9 @@ abstract type AbstractMultilayerUGraph{T,U} <: AbstractMultilayerGraph{T,U} end
 
 Add MultilayerVertex `V` to multilayer graph `mg`, provided that `node(V)` is a `Node` of `mg`. Return true if succeeds.
 """
-function add_vertex_specialized!(
+@traitfn function add_vertex_specialized!(
     mg::M, V::MultilayerVertex
-) where {T,U,M<:AbstractMultilayerUGraph{T,U}}
+) where {T,U, M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
     !has_node(mg, node(V)) && return false
     has_vertex(mg, V) && return false
 
@@ -42,7 +42,7 @@ end
 
 Remove [MultilayerVertex](@ref) `mv` from `mg`. Return true if succeeds, false otherwise.
 """
-function rem_vertex!(mg::AbstractMultilayerUGraph, V::MultilayerVertex)
+@traitfn function Graphs.rem_vertex!(mg::M, V::MultilayerVertex) where {M <: AbstractMultilayerGraph; !IsDirected{M}}
     # Check that the node exists and then that the vertex exists
     has_node(mg, V.node) || return false
     has_vertex(mg, V) || return false
@@ -95,7 +95,7 @@ end
 
 Return true if `mg` has edge between the `src` and `dst` (does not check edge or vertex metadata).
 """
-function has_edge(mg::M, src::T, dst::T) where {T,M<:AbstractMultilayerUGraph{T}}
+@traitfn function Graphs.has_edge(mg::M, src::T, dst::T) where {T,M<:AbstractMultilayerGraph{T}; !IsDirected{M}}
     has_vertex(mg, src) || return false
     has_vertex(mg, dst) || return false
 
@@ -114,11 +114,11 @@ end
 
 # Overloads that make AbstractMultilayerUGraph an extension of Graphs.jl. These are all well-inferred .
 """
-    edges(mg::M) where {T,U,M<:AbstractMultilayerUGraph{T,U}}
+    edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
 
 Return an list of all the edges of `mg`.
 """
-function edges(mg::M) where {T,U,M<:AbstractMultilayerUGraph{T,U}}
+@traitfn function Graphs.edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
     edge_list = MultilayerEdge{U}[]
 
     for (_src_v, halfedges) in enumerate(mg.fadjlist)
@@ -145,9 +145,9 @@ end
 
 Add MultilayerEdge `me` to the AbstractMultilayerUGraph `mg`. Return true if succeeds, false otherwise.
 """
-function add_edge_undirected!(
+@traitfn function add_edge_undirected!(
     mg::M, me::E
-) where {T,U,M<:AbstractMultilayerUGraph{T,U},E<:MultilayerEdge{<:Union{U,Nothing}}}
+) where {T,U,E<:MultilayerEdge{<:Union{U,Nothing}},M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
     _src = get_bare_mv(src(me))
     _dst = get_bare_mv(dst(me))
     has_vertex(mg, _src) ||
@@ -186,9 +186,9 @@ end
 
 Remove edge from `src` to `dst` from `mg`. Return true if succeeds, false otherwise.
 """
-function rem_edge_undirected!(
-    mg::AbstractMultilayerUGraph, src::MultilayerVertex, dst::MultilayerVertex
-)
+@traitfn function rem_edge_undirected!(
+    mg::M, src::MultilayerVertex, dst::MultilayerVertex
[…]
