+++ proposal/presentation/presentation.jl
@@ -1,0 +1,160 @@
+# Import the package manager 
+using Pkg
+
+# Install the package  
+Pkg.add("MultilayerGraphs")
+
+# Import the package 
+using MultilayerGraphs
+
+# Import necessary dependencies
+using Distributions, Graphs, SimpleValueGraphs
+
+# Set the number of nodes
+n_nodes = 100 
+
+# Create a list of nodes
+node_list = [Node("node_$i") for i in 1:n_nodes]
+
+# Create a simple directed layer
+n_vertices = rand(1:100)                          # Number of vertices 
+layer_simple_directed = layer_simpledigraph(      # Layer constructor 
+    :layer_simple_directed,                       # Layer name
+    sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
+    Truncated(Normal(5, 5), 0, 20),               # Indegree sequence distribution 
+    Truncated(Normal(5, 5), 0, 20)                # Outdegree sequence distribution
+)
+
+# Create a simple directed weighted layer
+n_vertices = rand(1:n_nodes)                                   # Number of vertices 
+n_edges = rand(n_vertices:(n_vertices * (n_vertices - 1) - 1)) # Number of edges 
+layer_simple_directed_weighted = layer_simpleweighteddigraph(  # Layer constructor 
+    :layer_simple_directed_weighted,                           # Layer name
+    sample(node_list, n_vertices; replace=false),              # Nodes represented in the layer
+    n_edges;                                                   # Number of randomly distributed edges
+    default_edge_weight=(src, dst) -> rand()                   # Function assigning weights to edges 
+)
+
+# Create a simple directed value layer
+n_vertices = rand(1:n_nodes)                                   # Number of vertices 
+n_edges = rand(n_vertices:(n_vertices * (n_vertices - 1) - 1)) # Number of edges 
+default_vertex_metadata = v -> ("vertex_$(v)_metadata")        # Vertex metadata 
+default_edge_metadata = (s, d) -> (rand(),)                    # Edge metadata 
+layer_simple_directed_value = Layer(                           # Layer constructor
+    :layer_simple_directed_value,                              # Layer name
+    sample(node_list, n_vertices; replace=false),              # Nodes represented in the layer
+    n_edges,                                                   # Number of randomly distributed edges
+    ValDiGraph(                                                
+        SimpleDiGraph{Int64}(); 
+        vertexval_types=(String,),
+        vertexval_init=default_vertex_metadata,
+        edgeval_types=(Float64,),
+        edgeval_init=default_edge_metadata,
+    ),
+    Float64;
+    default_vertex_metadata=default_vertex_metadata,           # Vertex metadata 
+    default_edge_metadata=default_edge_metadata                # Edge metadata 
+);
+
+# Create a list of layers 
+layers = [layer_simple_directed, layer_simple_directed_weighted, layer_simple_directed_value]
+
+# Create a simple directed interlayer
+n_vertices_1 = nv(layer_simple_directed)               # Number of vertices of layer 1
+n_vertices_2 = nv(layer_simple_directed_weighted)      # Number of vertices of layer 2
+n_edges = rand(1:(n_vertices_1 * n_vertices_2 - 1))    # Number of interlayer edges 
+interlayer_simple_directed = interlayer_simpledigraph( # Interlayer constructor 
+    layer_simple_directed,                             # Layer 1 
+    layer_simple_directed_weighted,                    # Layer 2 
+    n_edges                                            # Number of edges 
+)
+
+# Create a simple directed meta interlayer 
+n_vertices_1 = nv(layer_simple_directed_weighted)   # Number of vertices of layer 1
[…]
