+++ src/graphs_extensions/metagraphs.jl
@@ -84,6 +84,6 @@ function _set_metadata!(
     return set_props!(g, src, dst, Dict(key => value for (key, value) in pairs(metadata)))
 end
 
-weights(g::AbstractMetaGraph{T}) where {T} = adjacency_matrix(g)
+# weights(g::AbstractMetaGraph{T}) where {T} = adjacency_matrix(g)
 
 # is_directed(g::G) where {G <: AbstractMetaGraph} = is_directed(G)
