+++ src/graphs_extensions/simpleweightedgraphs.jl
@@ -1,4 +1,4 @@
-weights(g::G) where {T,G<:AbstractSimpleWeightedGraph{T}} = weights(g)
+# weights(g::G) where {T,G<:AbstractSimpleWeightedGraph{T}} = weights(g)
 
 function __add_vertex!(
     g::AbstractSimpleWeightedGraph{T}; metadata::Union{Tuple,NamedTuple}=NamedTuple()
