+++ proposal/abstract/abstract.md
@@ -27,4 +27,6 @@ MultilayerGraphs.jl has been integrated within the [**JuliaGraphs**](https://git
 - the extension of [Graphs.jl](https://github.com/JuliaGraphs/Graphs.jl) with several methods and metrics including the multilayer eigenvector centrality, the multilayer modularity and the Von Newman entropy; 
 - the compatibility with [Agents.jl](https://github.com/JuliaDynamics/Agents.jl) allowing for agent-based modelling on general multilayer networks. 
 
-For a comprehensive exploration of the package features and functionalities the readers is invited to consult the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md) and [documentation](https://juliagraphs.org/MultilayerGraphs).
+In our talk we will briefly introduce the theory and applications of multilayer graphs, showcase some of the main features of the current version of the package through a quick tutorial and talk about its planned future developments.
+
+For a more comprehensive exploration of the package functionalities and further details on the future developments the user is invited to consult the package [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [documentation](https://juliagraphs.org/MultilayerGraphs) and [issues](https://github.com/JuliaGraphs/MultilayerGraphs.jl/issues).
