+++ proposal/abstract/abstract.txt
@@ -13,7 +13,7 @@ and the JuliaDynamics [3] ecosystems through:
 
 In our talk we will briefly introduce the theory and applications of multilayer graphs, showcase some of the main features of the current version of the package through a quick tutorial and talk about its planned future developments.
 
-For a more comprehensive exploration of the package functionalities and further details on the future developments the user is invited to consult the package README [6], documentation [7] and issues [8].
+For a more comprehensive exploration of the package functionalities and further details on the future developments the reader is invited to consult the package README [6], documentation [7] and issues [8].
 
 KEYWORDS: Discrete Mathematics; Graph Theory; Network Science; Multilayer Graphs; Multilayer Networks; Complex Systems; Computer Science; Julia Language.
 
