### 2023-01-18 2915992
Restructure API, begin implementation of synchronized edge colored multilayer graph
Co-Authored-By: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-Authored-By: Claudio Moroni <43729990+ClaudMor@users.noreply.github.com>
