### 2023-01-19 ff922bc
Update
1. First complete implementation of synchronized edge colored graphs

2. Working sub-ecosystem prototype mainly based on abstract types. Traits are only used to discern between directed and undirercted graphs.

TODO:

Test the synchronized edge colored graphs

Co-Authored-By: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-Authored-By: Claudio Moroni <43729990+ClaudMor@users.noreply.github.com>
