### 2023-01-20 cf85df5
Update
Implementation of synchronized edge colored (di) graph completed

Implementation of sub-ecosystem endpoints completed

TODO
1. Update API.md
2. Update index.md
3. Update console printing of layers and interlayers
4. Add FAQ to readme.md & index.md
5. Clean up the code

Co-Authored-By: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-Authored-By: Claudio Moroni <43729990+ClaudMor@users.noreply.github.com>
