+++ src/undirected.jl
@@ -52,7 +52,7 @@ Remove [MultilayerVertex](@ref) `mv` from `mg`. Return true if succeeds, false o
 """
 @traitfn function _rem_vertex!(mg::M, V::MultilayerVertex) where {M <: AbstractMultilayerGraph; !IsDirected{M}}
     # Check that the node exists and then that the vertex exists
-    has_node(mg, V.node) || return false
+    has_node(mg, node(V)) || return false
     has_vertex(mg, V) || return false
 
     # Get the v corresponding to V, delete the association and replace it with a MissingVertex. Also substitute the metadata with an empty NamedTuple
