+++ src/multilayerdigraph.jl
@@ -322,117 +322,3 @@ Specify the interlayer `new_interlayer` as part of `mg`.
 
     return _specify_interlayer!(mg, new_interlayer;)
 end
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-#= """
-    add_edge_specialized!(mg::M, me::E) where {T,U, M <: MultilayerDiGraph{T,U}, E <: MultilayerEdge{ <: Union{U,Nothing}}}
-
-Add MultilayerEdge `me` to the MultilayerDiGraph `mg`. Return true if succeeds, false otherwise.
-"""
-add_edge_specialized!(mg::M, me::E) where {T,U,M<:MultilayerDiGraph{T,U},E<:MultilayerEdge{<:Union{U,Nothing}}} = add_edge_directed!(mg, me)
-
-
-"""
-    rem_edge_specialized!(mg::MultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex)
-
-Remove edge from `src` to `dst` from `mg`. Return true if succeeds, false otherwise.
-"""
-rem_edge_specialized!(mg::MultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex)  = rem_edge_directed!(mg, src, dst) =#
-
-
-#= """
-    rem_edge!(mg::MultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex)
-
-Remove edge from `src` to `dst` from `mg`. Return true if succeeds, false otherwise.
-"""
-function rem_edge!(
-    mg::MultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex
-)
-
-    # Perform routine checks
-    has_vertex(mg, src) ||
-        throw(ErrorException("Vertex $_src does not belong to the multilayer graph."))
-    has_vertex(mg, dst) ||
-        throw(ErrorException("Vertex $_dst does not belong to the multilayer graph."))
-
-    has_edge(mg, src, dst) || return false
-
-    src_V_idx = get_v(mg, src)
-    dst_V_idx = get_v(mg, dst)
-
-    _src = get_bare_mv(src)
-    _dst = get_bare_mv(dst)
-
-    src_idx_tbr = findfirst(halfedge -> vertex(halfedge) == _dst, mg.fadjlist[src_V_idx])
-    deleteat!(mg.fadjlist[src_V_idx], src_idx_tbr)
-
-    dst_idx_tbr = findfirst(halfedge -> halfedge.vertex == _src, mg.badjlist[dst_V_idx])
-    deleteat!(mg.badjlist[dst_V_idx], dst_idx_tbr)
-
-    return true
-end =#
-
-# Layers and Interlayers
-
-# Graphs.jl's extensions
-
-# Multilayer-specific methods
-# "empty graph" could be the correct way of calling a graph with no edges: https://math.stackexchange.com/questions/320859/what-is-the-term-for-a-graph-on-n-vertices-with-no-edges
-
-#= # Base overloads
-"""
-    Base.getproperty(mg::M, f::Symbol) where { M <: MultilayerDiGraph }
-"""
-function Base.getproperty(mg::MultilayerDiGraph, f::Symbol) # where {T,U,M<:AbstractMultilayerDiGraph{T,U}}
-    if f in (
-        :v_V_associations,
-        :fadjlist,
-        :badjlist,
-        :idx_N_associations,
-        :layers,
-        :interlayers,
-        :v_metadata_dict,
-    ) # :weight_tensor, :supra_weight_matrix, 
-        Base.getfield(mg, f)
-    elseif f == :edge_list
-        return edges(mg)
-    elseif f == :subgraphs
[…]
