### 2023-01-22 1fc488f
Eliminate multiplex graphs (moved to multiplex_graphs branch)
Co-Authored-By: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-Authored-By: Claudio Moroni <43729990+ClaudMor@users.noreply.github.com>
