+++ src/multiplexdigraph.jl
@@ -1,115 +1,0 @@
-
-"""
-MultiplexDiGraph{T, U, G <: AbstractGraph{T}} <: AbstractMultilayerGraph{T,U}
-
-A concrete type that can represent a general (directed) multiplex graph. Differently from `MultilayerDiGraph`, this type will error if:
-1. The user tries to specify one of its interlayers;
-2. The user tries to add and inter-layer edge which is not diagonal;
-3. The user tries to remove an inter-layer edge.
-
-Its internal fields aren't meant to be modified by the user. Please prefer the provided API.
-"""
-mutable struct MultiplexDiGraph{T,U} <: AbstractMultiplexGraph{T,U}
-layers::Vector{LayerDescriptor{T,U}} # vector containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.
-interlayers::OrderedDict{Set{Symbol},InterlayerDescriptor{T,U}} # the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
-v_V_associations::Bijection{T,<:MultilayerVertex} # A Bijection from Bijections.jl that associates numeric vertices to `MultilayerVertex`s.
-idx_N_associations::Bijection{Int64,Node} # A Bijection from Bijections.jl that associates Int64 to `Node`s.
-fadjlist::Vector{Vector{HalfEdge{<:MultilayerVertex,<:Union{Nothing,U}}}} # the forward adjacency list of the MultiplexDiGraph. It is a vector of vectors of `HalfEdge`s. Its i-th element are the `HalfEdge`s that originate from `v_V_associations[i]`.
-v_metadata_dict::Dict{T,<:Union{<:Tuple,<:NamedTuple}} #  A Dictionary that associates numeric vertices to their metadata.
-end
-
-# Traits
-@traitimpl IsWeighted{MultiplexDiGraph}
-@traitimpl IsMeta{MultiplexDiGraph}
-
-# Constructors
-"""
-MultiplexDiGraph(
-    layers::Vector{<:Layer{T,U}};
-    default_interlayers_null_graph::H = SimpleGraph{T}().
-) where {T,U, H <: AbstractGraph{T}}   
-
-Construct an undirected multiplex graph. The interlayers will have underlying graph type `default_interlayers_null_graph` and of course have only diagonal couplings.
-
-# ARGUMENTS
-
-- `layers::Vector{<:Layer{T,U}}`: The (ordered) list of layers the multilayer graph will have;
-- `default_interlayers_null_graph::H = SimpleGraph{T}()`: Sets the underlying graph for the interlayers that are to be automatically specified. Defaults to `SimpleGraph{T}()`. See the `Interlayer` constructors for more information;
-"""
-function MultiplexDiGraph(
-layers::Vector{<:Layer{T,U}};
-default_interlayers_null_graph::H=SimpleGraph{T}()
-) where {T,U,H<:AbstractGraph{T}}
-multilplexdigraph = MultiplexDiGraph(T, U)
-
-for layer in deepcopy(layers)
-    add_layer!(
-        multilplexdigraph,
-        layer;
-        default_interlayers_null_graph=default_interlayers_null_graph,
-        default_interlayers_structure="multiplex",
-    )
-end
-
-return multilplexdigraph
-end
-
-
-"""
-MultiplexDiGraph(T::Type{<:Number}, U::Type{<:Number})
-
[…]
