+++ src/directed.jl
@@ -9,8 +9,8 @@
 Add MultilayerVertex `V` to multilayer graph `mg`.  If `add_node` is true and `node(mv)` is not already part of `mg`, then add `node(mv)` to `mg` before adding `mv` to `mg` instead of throwing an error. Return true if succeeds. 
 """
 @traitfn function _add_vertex!(
-    mg::M, V::MultilayerVertex;  add_node::Bool=true
-) where {T,U, M<:AbstractMultilayerGraph{T,U}; IsDirected{M}}
+    mg::M, V::MultilayerVertex; add_node::Bool=true
+) where {T,U,M<:AbstractMultilayerGraph{T,U};IsDirected{M}}
     has_vertex(mg, V) && return false
     if add_node
         _node = node(V)
@@ -45,7 +45,9 @@ end
 
 Remove [MultilayerVertex](@ref) `mv` from `mg`. Return true if succeeds, false otherwise.
 """
-@traitfn function _rem_vertex!(mg::M, V::MultilayerVertex) where {M <: AbstractMultilayerGraph; IsDirected{M}}
+@traitfn function _rem_vertex!(
+    mg::M, V::MultilayerVertex
+) where {M <: AbstractMultilayerGraph; IsDirected{M}}
     # Check that the node exists and then that the vertex exists
     has_node(mg, V.node) || return false
     has_vertex(mg, V) || return false
@@ -118,7 +120,9 @@ end
 
 Return true if `mg` has edge between the `src` and `dst` (does not check edge or vertex metadata).
 """
-@traitfn function Graphs.has_edge(mg::M, src::T, dst::T) where {T,M<:AbstractMultilayerGraph{T}; IsDirected{M}}
+@traitfn function Graphs.has_edge(
+    mg::M, src::T, dst::T
+) where {T,M<:AbstractMultilayerGraph{T};IsDirected{M}}
     # Returns true if src is a vertex of mg
     has_vertex(mg, src) || return false
     # Returns true if dst is a vertex of mg
@@ -136,14 +140,15 @@ Return true if `mg` has edge between the `src` and `dst` (does not check edge or
     end
 end
 
-
 # Overloads that make AbstractMultilayerDiGraph an extension of Graphs.jl. These are all well-inferred.
 """
     edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; IsDirected{M}}
 
 Return an list of all the edges of `mg`.
 """
-@traitfn function Graphs.edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; IsDirected{M}}
+@traitfn function Graphs.edges(
+    mg::M
+) where {T,U,M<:AbstractMultilayerGraph{T,U};IsDirected{M}}
     edge_list = MultilayerEdge{U}[]
 
     for (_src_v, halfedges) in enumerate(mg.fadjlist)
@@ -161,7 +166,6 @@ Return an list of all the edges of `mg`.
     return edge_list
 end
 
-
 """
     _add_edge!(mg::M, me::E) where {T,U,E<:MultilayerEdge{<:Union{U,Nothing}},M<:AbstractMultilayerGraph{T,U}; IsDirected{M}}
 
@@ -169,7 +173,9 @@ Add MultilayerEdge `me` to the AbstractMultilayerDiGraph `mg`. Return true if su
 """
 @traitfn function _add_edge!(
     mg::M, me::E
-) where {T,U,E<:MultilayerEdge{<:Union{U,Nothing}},M<:AbstractMultilayerGraph{T,U}; IsDirected{M}}
+) where {
+    T,U,E<:MultilayerEdge{<:Union{U,Nothing}},M<:AbstractMultilayerGraph{T,U};IsDirected{M}
+}
     _src = get_bare_mv(src(me))
[…]
