+++ src/undirected.jl
@@ -7,8 +7,8 @@
 Add MultilayerVertex `V` to multilayer graph `mg`.  If `add_node` is true and `node(mv)` is not already part of `mg`, then add `node(mv)` to `mg` before adding `mv` to `mg` instead of throwing an error. Return true if succeeds. 
 """
 @traitfn function _add_vertex!(
-    mg::M, V::MultilayerVertex;  add_node::Bool=true
-) where {T,U, M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
+    mg::M, V::MultilayerVertex; add_node::Bool=true
+) where {T,U,M<:AbstractMultilayerGraph{T,U};!IsDirected{M}}
     has_vertex(mg, V) && return false
     if add_node
         _node = node(V)
@@ -18,7 +18,6 @@ Add MultilayerVertex `V` to multilayer graph `mg`.  If `add_node` is true and `n
     else
         !has_node(mg, node(V)) && return false
     end
-    
 
     n_nodes = nn(mg)
 
@@ -43,7 +42,9 @@ end
 
 Remove [MultilayerVertex](@ref) `mv` from `mg`. Return true if succeeds, false otherwise.
 """
-@traitfn function _rem_vertex!(mg::M, V::MultilayerVertex) where {M <: AbstractMultilayerGraph; !IsDirected{M}}
+@traitfn function _rem_vertex!(
+    mg::M, V::MultilayerVertex
+) where {M <: AbstractMultilayerGraph; !IsDirected{M}}
     # Check that the node exists and then that the vertex exists
     has_node(mg, node(V)) || return false
     has_vertex(mg, V) || return false
@@ -96,7 +97,9 @@ end
 
 Return true if `mg` has edge between the `src` and `dst` (does not check edge or vertex metadata).
 """
-@traitfn function Graphs.has_edge(mg::M, src::T, dst::T) where {T,M<:AbstractMultilayerGraph{T}; !IsDirected{M}}
+@traitfn function Graphs.has_edge(
+    mg::M, src::T, dst::T
+) where {T,M<:AbstractMultilayerGraph{T};!IsDirected{M}}
     has_vertex(mg, src) || return false
     has_vertex(mg, dst) || return false
 
@@ -112,14 +115,15 @@ Return true if `mg` has edge between the `src` and `dst` (does not check edge or
     end
 end
 
-
 # Overloads that make AbstractMultilayerUGraph an extension of Graphs.jl. These are all well-inferred .
 """
     edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
 
 Return an list of all the edges of `mg`.
 """
-@traitfn function Graphs.edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; !IsDirected{M}}
+@traitfn function Graphs.edges(
+    mg::M
+) where {T,U,M<:AbstractMultilayerGraph{T,U};!IsDirected{M}}
     edge_list = MultilayerEdge{U}[]
 
     for (_src_v, halfedges) in enumerate(mg.fadjlist)
@@ -140,7 +144,6 @@ Return an list of all the edges of `mg`.
     return unique(edge_list)
 end
 
-
 """
     _add_edge!(mg::M, me::E) where {T,U, M <: AbstractMultilayerUGraph{T,U}, E <: MultilayerEdge{ <: Union{U,Nothing}}}
 
@@ -148,7 +151,9 @@ Add MultilayerEdge `me` to the AbstractMultilayerUGraph `mg`. Return true if suc
 """
 @traitfn function _add_edge!(
     mg::M, me::E
[…]
