+++ src/graphs_extensions/simplevaluegraphs.jl
@@ -82,5 +82,4 @@ end
 
 # weights(g::SimpleValueGraphs.AbstractValGraph{T}) where {T} = adjacency_matrix(g)
 
-
 # is_directed(g::G) where {G <: SimpleValueGraphs.AbstractValGraph} = is_directed(G)

[…]
