+++ docs/make.jl
@@ -5,12 +5,12 @@ using Graphs
 using MultilayerGraphs
 using Documenter
 
-function end_user_filter(f) 
+#= function end_user_filter(f) 
     if f isa Function
         for method in methods
             tv, decls, file, line = arg_decl_parts(m)
             
-end
+end =#
 
 DocMeta.setdocmeta!(
     MultilayerGraphs, :DocTestSetup, :(using MultilayerGraphs); recursive=true
