+++ docs/src/API.md
@@ -5,5 +5,4 @@ This page provides a list of exported methods.
 ```@autodocs
 Modules = [MultilayerGraphs]
 Private = false
-Filter = myfilter
 ```
