+++ docs/src/API_organized.md
@@ -246,12 +246,12 @@ mv_inneighbors(mg::AbstractMultilayerGraph, mv::MultilayerVertex)
 mv_outneighbors(mg::AbstractMultilayerGraph, mv::MultilayerVertex)
 mv_neighbors( mg::AbstractMultilayerGraph, mv::MultilayerVertex)
 <!-- add_vertex!(mg::AbstractMultilayerGraph, mv::MultilayerVertex; add_node::Bool) -->
-rem_vertex!(mg::AbstractMultilayerUGraph, V::MultilayerVertex)
+<!-- rem_vertex!(mg::AbstractMultilayerUGraph, V::MultilayerVertex) -->
 <!-- rem_vertex!(mg::AbstractMultilayerDiGraph, V::MultilayerVertex) -->
 has_edge(mg::AbstractMultilayerGraph, edge::MultilayerEdge) 
 has_edge( subgraph::AbstractMultilayerGraph, s::MultilayerVertex, d::MultilayerVertex)
 ne(mg::AbstractMultilayerGraph)
-edges(mg::AbstractMultilayerUGraph)
+<!-- edges(mg::AbstractMultilayerUGraph) -->
 <!-- edges(mg::M) where {T,U,M<:AbstractMultilayerUGraph{T,U}}
 edges(mg::M) where {T,U,M<:AbstractMultilayerGraph{T,U}; IsDirected{M}} -->
 edges
@@ -266,25 +266,25 @@ get_weight(mg::AbstractMultilayerGraph, src::MultilayerVertex, dst::MultilayerVe
 <!-- set_weight!(mg::M, src::MultilayerVertex, dst::MultilayerVertex, weight::U) where { T,U, M <: AbstractMultilayerUGraph{T,U}}
 set_weight!(mg::M, src::MultilayerVertex, dst::MultilayerVertex, weight::U) where {T,U,M<:AbstractMultilayerGraph{T,U}; IsDirected{M}} -->
 set_weight!
-is_directed(mg::AbstractMultilayerUGraph)
-is_directed(m::M) where { M <: Type{ <: AbstractMultilayerUGraph}}
-is_directed(mg::AbstractMultilayerDiGraph)
-is_directed(m::M) where { M <: Type{ <: AbstractMultilayerDiGraph}}
+<!-- is_directed(mg::AbstractMultilayerUGraph) -->
+<!-- is_directed(m::M) where { M <: Type{ <: AbstractMultilayerUGraph}} -->
+<!-- is_directed(mg::AbstractMultilayerDiGraph)
+is_directed(m::M) where { M <: Type{ <: AbstractMultilayerDiGraph}} -->
 has_node(mg::AbstractMultilayerGraph, n::Node)
 <!-- set_metadata!(mg::AbstractMultilayerGraph, mv::MultilayerVertex, metadata::Union{Tuple, NamedTuple}) 
-set_metadata!(mg::AbstractMultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex, metadata::Union{Tuple, NamedTuple})
+<!-- set_metadata!(mg::AbstractMultilayerDiGraph, src::MultilayerVertex, dst::MultilayerVertex, metadata::Union{Tuple, NamedTuple}) -->
 set_metadata!(mg::AbstractMultilayerUGraph, src::MultilayerVertex, dst::MultilayerVertex, metadata::Union{Tuple, NamedTuple}) -->
 set_metadata!
 nl(mg::AbstractMultilayerGraph)
 nIn(mg::AbstractMultilayerGraph)
 has_layer(mg::AbstractMultilayerGraph, layer_name::Symbol)
[…]
