### 2023-01-24 364692e
Traits based sub ecosystem & `NodeAlingnedEdgeColored(Di)Graph` (#107)
