+++ src/multilayergraph.jl
@@ -3,7 +3,7 @@
 
 A concrete type that can represent a general multilayer graph. Its internal fields aren't meant to be modified by the user. Please prefer the provided API.
 """
-mutable struct MultilayerGraph{T,U} <: AbstractMultilayerUGraph{T,U}
+mutable struct MultilayerGraph{T,U} <: AbstractMultilayerGraph{T,U}
     layers::Vector{LayerDescriptor{T,U}} # vector containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.
     interlayers::OrderedDict{Set{Symbol},InterlayerDescriptor{T,U}} # the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
     v_V_associations::Bijection{T,<:MultilayerVertex} # A Bijection from Bijections.jl that associates numeric vertices to `MultilayerVertex`s.
@@ -14,6 +14,14 @@ end
 
 # Traits
 @traitimpl IsWeighted{MultilayerGraph}
+@traitimpl IsMeta{MultilayerGraph}
+
+"""
+    is_directed(m::M) where { M <: Type{ <: MultilayerGraph}}
+
+Return `false`
+"""
+Graphs.is_directed(mg::M) where {M<:Type{<:MultilayerGraph}} = false
 
 # Constructors
 """
@@ -30,7 +38,7 @@ Construct a MultilayerGraph with layers given by `layers`. The interlayers will 
 
 - `layers::Vector{<:Layer{T,U}}`: The (ordered) list of layers the multilayer graph will have;
 - `specified_interlayers::Vector{<:Interlayer{T,U}}`: The list of interlayers specified by the user. Note that the user does not need to specify all interlayers, as the unspecified ones will be automatically constructed using the indications given by the `default_interlayers_null_graph` and `default_interlayers_structure` keywords;
-- `default_interlayers_null_graph::H = SimpleGraph{T}()`: Sets the underlying graph for the interlayers that are to be automatically specified. Defaults to `SimpleGraph{T}()`. See the `Layer` constructors for more information;
+- `default_interlayers_null_graph::H = SimpleGraph{T}()`: Sets the underlying graph for the interlayers that are to be automatically specified. Defaults to `SimpleGraph{T}()`. See the `Interlayer` constructors for more information;
 - `default_interlayers_structure::String = "multiplex"`: Sets the structure of the interlayers that are to be automatically specified. May be "multiplex" for diagonally coupled interlayers, or "empty" for empty interlayers (no edges).  "multiplex". See the `Interlayer` constructors for more information.
 """
 function MultilayerGraph(
@@ -187,139 +195,124 @@ function MultilayerGraph(
     )
 end
 
-# General MultilayerGraph Utilities
-fadjlist(mg::MultilayerGraph) = mg.fadjlist
-
 # Nodes
 
+"""
+    add_node!(mg::MultilayerGraph, n::Node; add_vertex_to_layers::Union{Vector{Symbol}, Symbol} = Symbol[])
+
+Add node `n` to `mg`. Return true if succeeds. Additionally, add a corresponding vertex to all layers whose name is listed in `add_vertex_to_layers`. If `add_vertex_to_layers == :all`, then a corresponding vertex is added to all layers.
+"""
+function add_node!(
+    mg::MultilayerGraph,
+    n::Node;
[…]
