+++ docs/src/index.md
@@ -939,6 +939,14 @@ Similarly, there is a [`MetadataTensor`](@ref), that may be created via `metadat
 
 The package also exports a [`SupraWeightMatrix`](@ref) which is a supra (weighted) adjacency matrix with the same indexing functionality as above. You may instantiate it via `supra_weight_matrix(multilayergraph)`.
 
+
+#### Sub-ecosystem
+Special applications may not require all the representational generality enabled by `Multilayer(Di)Graph`s, and, on the contrary, could benefit from the simpler interface and higher performance that come with restricted subtypes of multilayer graphs (e.g. multiplex graphs, edge-colored graph, etc).
+
+MultilayerGraphs.jl, via an apporoach that combines type-hierarchy with traits, allows for implementing custom multilayer graphs (similar to what Graphs.jl does). This feature has been initially proven with the implementation of [`SynchronizedEdgeColoredGraph`](@ref) and [`SynchronizedEdgeColoredDiGraph`](@ref), which aim at representing edge-colored graphs by naturally mapping them to multilayer graphs.
+
+
+
 #### Multilayer-specific analytical tools
 
 Read a complete list of analytical methods exclusive to multilayer graphs in the dedicated [API section](@ref msm_eu) (here "exclusive" means that wither those methods do not exists for standard graphs, or that they had to be reimplemented and so may present some caveats). Refer to their docstrings for more information.

[…]
