+++ docs/make.jl
@@ -26,7 +26,7 @@ makedocs(;
         canonical="https://juliagraphs.org/MultilayerGraphs.jl",
         assets=String[],
     ),
-    pages=["Home" => "index.md", "API" => "API.md"],
+    pages=["Home" => "index.md", "API" => "API_organized.md"],
     clean=false,
 )
 
