+++ docs/src/API_autodocs.md
@@ -1,8 +1,0 @@
-# API
-
-This page provides a list of exported methods.
-
-```@autodocs
-Modules = [MultilayerGraphs]
-Private = false
-```
