### 2023-01-25 d7eab86
CompatHelper: add new compat entry for Graphs at version 1, (keep existing compat)
