+++ Project.toml
@@ -26,6 +26,7 @@ WhereTraits = "c9d4e05b-6318-49cb-9b56-e0e2b0ceadd8"
 Bijections = "0.1"
 DataStructures = "0.18"
 Distributions = "0.25"
+Graphs = "1"
 IterTools = "1.4"
 MetaGraphs = "0.7"
 OMEinsum = "0.7"
