### 2023-01-25 76347bf
CompatHelper: add new compat entry for WhereTraits at version 1, (keep existing compat)
