+++ Project.toml
@@ -34,4 +34,5 @@ SimpleTraits = "0.9"
 SimpleValueGraphs = "0.4"
 SimpleWeightedGraphs = "1.2"
 TensorOperations = "3.2"
+WhereTraits = "1"
 julia = "1.8"
