+++ .github/workflows/CI.yml
@@ -18,8 +18,8 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.8.2'
-          - 'nightly'
+          - '1'
+          #- 'nightly'
         os:
           - ubuntu-latest
           - macOS-latest
