+++ .github/workflows/CI.yml
@@ -19,6 +19,7 @@ jobs:
       matrix:
         version:
           - '1'
+          - '1.8'
           #- 'nightly'
         os:
           - ubuntu-latest
