+++ src/multilayerdigraph.jl
@@ -279,11 +279,18 @@ end
 # Layers and Interlayers
 """
     add_layer!(
-        mg::M, 
-        new_layer::L; 
-        default_interlayers_null_graph::H = SimpleGraph{T}(), 
-        default_interlayers_structure::String ="multiplex"
-    ) where {T,U,G<:AbstractGraph{T},M<:MultilayerDiGraph{T,U},L<:Layer{T,U,G}, H <: AbstractGraph{T}}
+        mg::M,
+        new_layer::L;
+        default_interlayers_null_graph::H=SimpleGraph{T}(),
+        default_interlayers_structure::String="multiplex",
+    ) where {
+        T,
+        U,
+        G<:AbstractGraph{T},
+        L<:Layer{T,U,G},
+        H<:AbstractGraph{T},
+        M<:MultilayerDiGraph{T,U}
+    }
     
 Add layer `layer` to `mg`.
 
@@ -294,7 +301,7 @@ Add layer `layer` to `mg`.
 - `default_interlayers_null_graph::H`: upon addition of a new `Layer`, all the `Interlayer`s between the new and the existing `Layer`s are immediately created. This keyword argument specifies their `null_graph` See the `Layer` constructor for more information. Defaults to `SimpleGraph{T}()`;
 - `default_interlayers_structure::String`: The structure of the `Interlayer`s created by default. May either be "multiplex" to have diagonally-coupled only interlayers, or "empty" for empty interlayers. Defaults to "multiplex".
 """
-@traitfn function add_layer!(
+function add_layer!(
     mg::M,
     new_layer::L;
     default_interlayers_null_graph::H=SimpleGraph{T}(),
@@ -305,7 +312,7 @@ Add layer `layer` to `mg`.
     G<:AbstractGraph{T},
     L<:Layer{T,U,G},
     H<:AbstractGraph{T},
-    M<:MultilayerDiGraph{T,U};IsDirected{M},
+    M<:MultilayerDiGraph{T,U}
 }
     return add_layer_directedness!(
         mg,
