+++ docs/src/index.md
@@ -943,9 +943,9 @@ The package also exports a [`SupraWeightMatrix`](@ref) which is a supra (weighte
 #### Sub-ecosystem
 Special applications may not require all the representational generality enabled by `Multilayer(Di)Graph`s, and, on the contrary, could benefit from the simpler interface and higher performance that come with restricted subtypes of multilayer graphs (e.g. multiplex graphs, edge-colored graph, etc).
 
-MultilayerGraphs.jl, via an apporoach that combines type-hierarchy with traits, allows for implementing custom multilayer graphs (similar to what Graphs.jl does). This feature has been initially proven with the implementation of [`SynchronizedEdgeColoredGraph`](@ref) and [`SynchronizedEdgeColoredDiGraph`](@ref), which aim at representing edge-colored graphs by naturally mapping them to multilayer graphs.
+MultilayerGraphs.jl, via an apporoach that combines type-hierarchy with traits, allows for implementing custom multilayer graphs (similar to what Graphs.jl does). This feature has been initially proven with the implementation of [`NodeAlignedEdgeColoredGraph`](@ref) and [`NodeAlignedEdgeColoredDiGraph`](@ref), which aim at representing edge-colored graphs by naturally mapping them to multilayer graphs.
 
-
+The sub-ecosystem capability is still under construction, see [Systematize the sub-ecosytem feature](https://github.com/JuliaGraphs/MultilayerGraphs.jl/issues/113).
 
 #### Multilayer-specific analytical tools
 
