+++ docs/src/API.md
@@ -1,6 +1,6 @@
 # API
 
-This page provides a list of exported methods organized by topic and audience. Methods that act on vertices, edges, and layers are grouped together. Some methods are intended for developers who want to use the `Graphs.jl` library as part of their code, while others are meant for end-users. hello
+This page provides a list of exported methods organized by topic and audience. Methods that act on vertices, edges, and layers are grouped together. Some methods are intended for developers who want to use the `Graphs.jl` library as part of their code, while others are meant for end-users.
 
 ## End-User
 
@@ -171,122 +171,44 @@ graph(subgraph::AbstractSubGraph)
 ### [Multilayer-Specific Methods](@id msm_eu)
 
 ```@docs
-MultilayerGraph{T,U}
+AbstractMultilayerGraph{T <: Integer, U <: Real}
 
-MultilayerDiGraph{T,U}
-
-MultilayerGraph(T::Type{<:Number}, U::Type{<:Number})
-
-MultilayerDiGraph(T::Type{<:Number}, U::Type{<:Number})
-
-MultilayerGraph(
-    layers::Vector{<:Layer{T,U}},
-    specified_interlayers::Vector{<:Interlayer{T,U}};
-    default_interlayers_null_graph::H = SimpleGraph{T}(),
-    default_interlayers_structure::String="multiplex",
-) where {T,U, H <: AbstractGraph{T}}
-
-MultilayerDiGraph(
-    layers::Vector{<:Layer{T,U}},
-    specified_interlayers::Vector{<:Interlayer{T,U}};
-    default_interlayers_null_graph::H = SimpleGraph{T}(),
-    default_interlayers_structure::String="multiplex",
-) where {T,U, H <: AbstractGraph{T}}
-
-MultilayerGraph(
-    empty_layers::Vector{<:Layer{T,U}},
-    empty_interlayers::Vector{<:Interlayer{T,U}},
-    degree_distribution::UnivariateDistribution;
-    allow_self_loops::Bool = false,
-    default_interlayers_null_graph::H = SimpleGraph{T}(),
-) where {T <: Integer, U <: Real, H <: AbstractGraph{T}}
-
-MultilayerDiGraph(
-    empty_layers::Vector{<:Layer{T,U}},
-    empty_interlayers::Vector{<:Interlayer{T,U}},
-    indegree_distribution::UnivariateDistribution,
-    outdegree_distribution::UnivariateDistribution;
-    allow_self_loops::Bool = false,
-    default_interlayers_null_graph::H = SimpleGraph{T}(),
-) where {T <: Integer, U <: Real, H <: AbstractGraph{T}}
-
-MultilayerGraph(
-    empty_multilayergraph::MultilayerGraph{T,U}, 
-    degree_sequence::Vector{<:Integer}; 
-    allow_self_loops::Bool = false, 
-    perform_checks::Bool = true
-) where {T,U}
-
-MultilayerDiGraph(
-    empty_multilayerdigraph::MultilayerDiGraph{T,U}, 
-    indegree_sequence::Vector{<:Integer},
-    outdegree_sequence::Vector{<:Integer};
-    allow_self_loops::Bool = false,
[…]
