+++ paper/paper.md
@@ -25,8 +25,8 @@ affiliations:
    index: 1
  - name: Interdisciplinary Physics Team, Italy
    index: 2
-date: 11 January 2023
-bibliography: paper.bib
+date: 27 January 2023
+bibliography: references.bib
 ---
 
 # Summary
