### 2023-01-30 #5116 @editorialbot
[REVIEW]: MultilayerGraphs.jl: Multilayer Network Science in Julia
**Submitting author:** @pitmonticone (<a href="http://orcid.org/0000-0002-2731-9623">Pietro Monticone</a>)
**Repository:** https://github.com/JuliaGraphs/MultilayerGraphs.jl
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 1.1.4
