### 2023-02-19 #5116 @TomKellyGenetics
Apologies for the slow follow-up due to unrelated personal reasons. I’m delighted to see that your software is well-documented and the paper is an excellent summary. The writing style differs from my field but it is easily understandable for a non-expert audience. I know from experience this is a difficult tasks that the authors have handled well. Excellent work.

Please give me more time to test the installation and run the code before giving more detailed feedback. So far I only have minor suggestions so if the quality of the software is at a similar level, I’d expect only few revisions would be necessary to give my recommendation to accept.
