### 2023-02-23 #5116 @TomKellyGenetics
I confirmed that all code examples in the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl#readme) execute without errors. The [Tutorial](https://juliagraphs.org/MultilayerGraphs.jl/stable/#Tutorial) however, give the following error when running some parts.

> ERROR: MethodError: no method matching Layer(::Symbol, ::Vector{Node}, ::Float64, ::SimpleGraph{Int64}, ::Type{Float64}

Please check it is up-to-date with the current version of the package and can be run without aliases on a new user's environment.

On the other hand, the Tutorial does a better job at demonstrating how to call created objects and check their contents. The REPL prints helpful information while executing commands but this would be enhanced by demonstrations on to integrate multilayer graphs with existing systems to print Graph.jl components or visualise the graphs created. Rather than avoiding redundancy with other packages, stand-alone documentation that does not assume users have familiarity with them would be ideal.

The authors should be commended for preparing high quality short-form documentation and manuscript which concisely summarises the importance of this contribution. Clearer long-form documentation is recommended but not essential. I'll note that the R packages for this theme do not have vignettes demonstrating this functionality either.

For comparison, the documentation for the Leiden clustering algorithm in Python and R (by Vincent Traag and myself respectively) has examples of multilayer graphs with visualisations.
https://leidenalg.readthedocs.io/en/latest/multiplex.html#layer-multiplex
https://cran.r-project.org/web/packages/leiden/vignettes/run_multiplex.html

As a suggestion to extend this work, Louvain or Leiden clustering would be compatible with this framework. It is a significant amount of work not required to publish in my opinion (but within the scope of issue 94).
