### 2023-02-23 #5116 @pitmonticone
Dear @TomKellyGenetics, thank you very much for your helpful comments.

### Installation

> While the configuration file automates installing other packages as dependencies, the software dependencies necessary for the user to understand should be clearly documented. Please list the minimum version of Julia (1.8) clearly in the installation instructions (e.g., under a 'System requirements' subheading). This should be a simple change to the documentation (and optionally the GitHub README).

We have changed the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [example](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.md) and [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/) accordingly. The stable documentation will be updated with the registration of the post-publication release.

### Usage Example

> Regarding the examples, my preference is to include them in the manuscript but I'll defer to editorial guidance and authorial discretion. However, I'll note this was already proposed in "pre-review" stage and the usage is described in the documentation linked: https://juliagraphs.org/MultilayerGraphs.jl/stable/

Yes, @diehlpk [suggested](https://github.com/openjournals/joss-reviews/issues/5055#issuecomment-1377805135) to create an example folder and to link it to the manuscript, so we did exactly that.

### Community Guidelines

> Regarding community guidelines, I note these are described briefly here (https://juliagraphs.org/MultilayerGraphs.jl/stable/#How-to-Contribute) but Issues or Pull Request templates would be beneficial to guide new contributors. See here for details on how to set it up: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository

We have added [issue templates](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/main/.github/ISSUE_TEMPLATE) accordingly.
