+++ src/multilayerdigraph.jl
@@ -307,12 +307,7 @@ function add_layer!(
     default_interlayers_null_graph::H=SimpleDiGraph{T}(),
     default_interlayers_structure::String="multiplex",
 ) where {
-    T,
-    U,
-    G<:AbstractGraph{T},
-    L<:Layer{T,U,G},
-    H<:AbstractGraph{T},
-    M<:MultilayerDiGraph{T,U}
+    T,U,G<:AbstractGraph{T},L<:Layer{T,U,G},H<:AbstractGraph{T},M<:MultilayerDiGraph{T,U}
 }
     return add_layer_directedness!(
         mg,
