+++ src/special_multilayergraphs/node_aligned_edge_colored/node_aligned_edge_colored_di_graph.jl
@@ -10,7 +10,7 @@ A concrete type that can represent a general directed edge colored graph, that i
 Its internal fields aren't meant to be modified by the user. Please prefer the provided API.
 """
 mutable struct NodeAlignedEdgeColoredDiGraph{T,U} <:
-        AbstractNodeAlignedEdgeColoredGraph{T,U}
+               AbstractNodeAlignedEdgeColoredGraph{T,U}
     layers::Vector{LayerDescriptor{T,U}} # vector containing all the layers of the multilayer graph. Their underlying graphs must be all undirected.
     interlayers::OrderedDict{Set{Symbol},InterlayerDescriptor{T,U}} #  the ordered dictionary containing all the interlayers of the multilayer graph. Their underlying graphs must be all undirected.
     v_V_associations::Bijection{T,<:MultilayerVertex} # A Bijection from Bijections.jl that associates numeric vertices to `MultilayerVertex`s.
