+++ proposal/presentation/presentation.jl
@@ -11,7 +11,7 @@ using MultilayerGraphs
 using Distributions, Graphs, SimpleValueGraphs
 
 # Set the number of nodes
-n_nodes = 100 
+n_nodes = 100
 
 # Create a list of nodes
 node_list = [Node("node_$i") for i in 1:n_nodes]
@@ -22,7 +22,7 @@ layer_simple_directed = layer_simpledigraph(      # Layer constructor
     :layer_simple_directed,                       # Layer name
     sample(node_list, n_vertices; replace=false), # Nodes represented in the layer
     Truncated(Normal(5, 5), 0, 20),               # Indegree sequence distribution 
-    Truncated(Normal(5, 5), 0, 20)                # Outdegree sequence distribution
+    Truncated(Normal(5, 5), 0, 20),                # Outdegree sequence distribution
 )
 
 # Create a simple directed weighted layer
@@ -32,7 +32,7 @@ layer_simple_directed_weighted = layer_simpleweighteddigraph(  # Layer construct
     :layer_simple_directed_weighted,                           # Layer name
     sample(node_list, n_vertices; replace=false),              # Nodes represented in the layer
     n_edges;                                                   # Number of randomly distributed edges
-    default_edge_weight=(src, dst) -> rand()                   # Function assigning weights to edges 
+    default_edge_weight=(src, dst) -> rand(),                   # Function assigning weights to edges 
 )
 
 # Create a simple directed value layer
@@ -44,8 +44,8 @@ layer_simple_directed_value = Layer(                           # Layer construct
     :layer_simple_directed_value,                              # Layer name
     sample(node_list, n_vertices; replace=false),              # Nodes represented in the layer
     n_edges,                                                   # Number of randomly distributed edges
-    ValDiGraph(                                                
-        SimpleDiGraph{Int64}(); 
+    ValDiGraph(
+        SimpleDiGraph{Int64}();
         vertexval_types=(String,),
         vertexval_init=default_vertex_metadata,
         edgeval_types=(Float64,),
@@ -53,11 +53,13 @@ layer_simple_directed_value = Layer(                           # Layer construct
     ),
     Float64;
     default_vertex_metadata=default_vertex_metadata,           # Vertex metadata 
-    default_edge_metadata=default_edge_metadata                # Edge metadata 
+    default_edge_metadata=default_edge_metadata,                # Edge metadata 
 );
 
 # Create a list of layers 
-layers = [layer_simple_directed, layer_simple_directed_weighted, layer_simple_directed_value]
+layers = [
+    layer_simple_directed, layer_simple_directed_weighted, layer_simple_directed_value
+]
 
 # Create a simple directed interlayer
 n_vertices_1 = nv(layer_simple_directed)               # Number of vertices of layer 1
@@ -66,7 +68,7 @@ n_edges = rand(1:(n_vertices_1 * n_vertices_2 - 1))    # Number of interlayer ed
 interlayer_simple_directed = interlayer_simpledigraph( # Interlayer constructor 
     layer_simple_directed,                             # Layer 1 
     layer_simple_directed_weighted,                    # Layer 2 
-    n_edges                                            # Number of edges 
+    n_edges,                                            # Number of edges 
 )
 
 # Create a simple directed meta interlayer 
@@ -78,8 +80,8 @@ interlayer_simple_directed_meta = interlayer_metadigraph( # Interlayer construct
     layer_simple_directed_value,                          # Layer 2
     n_edges;                                              # Number of edges
     default_edge_metadata=(src, dst) ->                   # Edge metadata 
[…]
