+++ test/runtests.jl
@@ -491,7 +491,7 @@ all_interlayers = [
     @testset "interlayer" begin
         include("interlayer.jl")
     end
-   
+
     @testset "undirected" begin
         include("undirected.jl")
     end
