### 2023-02-24 #5116 @bgailleton
`MultilayersGraphs.jl` is a Julia package extending `Graphs.jl` (a well-established package to create and manipulate all sorts of graph data structures) by adding to the numerical ecosystem Multilayer Graphs.

Overall, the package is well-documented (overview, installation, usage and API); the code itself well structured and reasonably well commented; and the manuscript clear, concise and understandable for the non-specialists.

I believe it represents a great addition to JOSS and reccomend publication, after maybe minor changes. I have just submitted an issue with more detailed comments and eventual reccomendations on the repository to keep this thread clear.
