### 2023-02-24 #5116 @ClaudMor
Dear @TomKellyGenetics ,

I'd like to integrate what's been written by @pitmonticone with two notes regarding the error you got while running the tutorial and the issue about console printing and visualizations.

## Error inside the Tutorial

>   I confirmed that all code examples in the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl#readme) execute without errors. The [Tutorial](https://juliagraphs.org/MultilayerGraphs.jl/stable/#Tutorial) however, give the following error when running some parts.
ERROR: MethodError: no method matching Layer(::Symbol, ::Vector{Node}, ::Float64, ::SimpleGraph{Int64}, ::Type{Float64}
Please check it is up-to-date with the current version of the package and can be run without aliases on a new user's environment.

I created a fresh environment where only essential packages have been installed:

```julia
Status `...\Project.toml`
  [31c24e10] Distributions v0.25.83
  [86223c79] Graphs v1.8.0
  [626554b9] MetaGraphs v0.7.2
  [af984530] MultilayerGraphs v1.1.3
  [b43c691f] SimpleValueGraphs v0.4.1
  [47aef6b3] SimpleWeightedGraphs v1.3.0
  [2913bbd2] StatsBase v0.33.21
```

And re-run the parts of the tutorial that instantiate `Layer`s. Although I ran them multiple times to ensure it was not coincidental, I could not reproduce the error. Could you point us to the line where it occurs and maybe also copy-paste the entire stacktrace? From the error message you reported above it seems that the "random constructor" of the `Layer` receives a real-valued number of edges, which makes me suspect it could be an error in the Tutorial rather than in the package, although I cannot find it at the moment.

While running the tutorial, I found an error on another line, which I corrected.

## Console print and Visualisations

> On the other hand, the Tutorial does a better job at demonstrating how to call created objects and check their contents. The REPL prints helpful information while executing commands but this would be enhanced by demonstrations on to integrate multilayer graphs with existing systems to print Graph.jl components or visualise the graphs created. Rather than avoiding redundancy with other packages, stand-alone documentation that does not assume users have familiarity with them would be ideal.

We're glad you found the displayed information useful. When a multilayer graph-object, a `Layer` or an `Interlayer` are entered in the REPL, a quick summary of their components is shown, including the underlying graph type(s). Thus, I'm not quite clear on what you mean by "this would be enhanced by demonstrations on to integrate multilayer graphs with existing systems to print Graph.jl components". On which occasions would you, more specifically, like to see more information displayed?

Regarding visualizations, the package does not have that functionality yet, but, just as the clustering algorithms mentioned above, they are part of the package's future plans.
