### 2023-02-24 JuliaGraphs/MultilayerGraphs.jl#129 @pitmonticone
Dear @bgailleton,

Thank you very much for your precious review comments.

### Installation

We have changed the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [example](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.md) and [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/) accordingly. The stable documentation will be updated with the registration of the post-publication release.
