### 2023-02-24 JuliaGraphs/MultilayerGraphs.jl#129 @pitmonticone
Dear @bgailleton,

Thank you very much for your precious review comments and suggestions.

### Installation

> Installation was straightforward, but I echo @TomKellyGenetics that there might be minor versioning conflicts. I managed to install it on Julia 1.8.6 (Ubuntu 22.04) but had to update Julia before. It is not a problem but the documentation should explicitly mention the minimum required version.

- [ ] Change the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [example](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.md) and [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/) accordingly. The stable documentation will be updated with the registration of the post-publication release.

### Usage Example

> The example file and tutorial in the documentation need extra dependencies. Julia is not my main language so my environment is rather empty and I had to install a bunch of additional packages (`Distributions, SimpleValueGraph, LoggingExtras, StatsBase, SimpleWeightedGraphs, MetaGraphs, Agents`). I recommend to update the documentation accordingly in order to save back-and-forth operations to install dependencies 1 by 1. This also regards the readme file in the example folder. The latter could be renamed `README.md` in order to automatically render when the folder is opened (I leave that to the authors' choice).

- [ ] Add the [environment configuration file]() and the necessary code to activate and instantiate that environment in order to ensure full reproducibility both in [example.jl]() and [example.md]().
- [ ] Update documentation code to install all necessary dependencies in [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/). The stable documentation will be updated with the registration of the post-publication release.
- [ ] Rename `example.md` into `README.md`
