### 2023-02-24 #5116 @pitmonticone
## Author Checklist

- [x] Change the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [example](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.md) and [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/) specifying the minimum version. The stable documentation will be updated with the registration of the post-publication release
- [x] Add [issue templates](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/main/.github/ISSUE_TEMPLATE)
- [x] Add the [environment configuration file](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/Project.toml) and the necessary code to activate and instantiate that environment in order to ensure full reproducibility both in [example.jl](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.jl) and [example.md](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/README.md).
- [x] Update documentation code to install all necessary dependencies in [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/). The stable documentation will be updated with the registration of the post-publication release.
- [x] Rename `example.md` into `README.md`
