### 2023-02-24 JuliaGraphs/MultilayerGraphs.jl#129 @pitmonticone
I've added an [author checklist](https://github.com/openjournals/joss-reviews/issues/5116#issuecomment-1443799043) in the main issue thread.
