### 2023-02-28 #5116 @TomKellyGenetics
The updated documentation is really helpful to understand if the code has created a graph object in the environment. I've installed the development version of the package from GitHub bearing in mind this will likely be released to accompany the publication (and direct users to cite it).

I checked my environment settings but I am still not able to run the tutorial examples. For instance:

```
julia> using MultilayerGraphs

julia> layer_sg = []
Any[]

julia> _nv, _ne  = rand_nv_ne_layer(min_vertices,max_vertices)
(6, 14.0)

julia> layer_sg = Layer(   :layer_sg,
                           sample(nodes_list, _nv, replace = false),
                           _ne,
                           SimpleGraph{vertextype}(),
                           _weighttype
       )
ERROR: MethodError: no method matching Layer(::Symbol, ::Vector{Node}, ::Float64, ::SimpleGraph{Int64}, ::Type{Float64})
Closest candidates are:
  Layer(::Symbol, ::Union{Vector{MultilayerVertex{nothing}}, Vector{Node}}, ::Vector{<:Integer}, ::G, ::Type{<:Real}; kwargs...) where G<:AbstractGraph at ~/.julia/packages/SimpleTraits/l1ZsK/src/SimpleTraits.jl:338
  Layer(::Symbol, ::Union{Vector{MultilayerVertex{nothing}}, Vector{Node}}, ::UnivariateDistribution, ::G, ::Type{U}; kwargs...) where {U<:Real, G<:AbstractGraph} at ~/.julia/packages/SimpleTraits/l1ZsK/src/SimpleTraits.jl:338
  Layer(::Symbol, ::Array{Union{V, N}, 1}, ::Int64, ::G, ::Type{U}; default_vertex_metadata, default_edge_weight, default_edge_metadata, allow_self_loops) where {T<:Integer, U<:Real, G<:AbstractGraph{T}, V<:MultilayerVertex{nothing}, N<:Node} at ~/.julia/packages/MultilayerGraphs/E0SGA/src/subgraphs/layer.jl:190
  ...
Stacktrace:
 [1] top-level scope
   @ REPL[57]:1

julia> VERSION
v"1.8.5"

julia> versioninfo()
Julia Version 1.8.5
Commit 17cfb8e65ea (2023-01-08 06:45 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: 24
  WORD_SIZE: 64
  Threads: 1 on 24 virtual cores
Environment:

julia> Pkg.status()
Status `~/.julia/environments/v1.8/Project.toml`
^ [31c24e10] Distributions v0.25.45
  [86223c79] Graphs v1.8.0
  [626554b9] MetaGraphs v0.7.2
  [af984530] MultilayerGraphs v1.1.3
  [6f49c342] RCall v0.13.14
  [295af30f] Revise v3.5.1
  [b43c691f] SimpleValueGraphs v0.4.1
  [47aef6b3] SimpleWeightedGraphs v1.3.0
  [2913bbd2] StatsBase v0.33.21
Info Packages marked with ^ have new versions available and may be upgradable.
```

There are similar issues with below examples as a result of this missing object. It's unclear if this is an issue with installing dependencies or the example in the documentation needing an update. Conditional on this example running without errors, I will recommend acceptance on the manuscript. Visualisation would be "nice to have" but I understand this may be under the scope of Graphs.jl or other related projects.
