### 2023-02-28 #5116 @pitmonticone
Hi @TomKellyGenetics,

Thank you for checking the reproducibility of the code chunks included in the [Tutorial](https://juliagraphs.org/MultilayerGraphs.jl/dev/#Tutorial) section of the package development documentation.

We've just added a [script](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/docs/src/tutorial.jl) containing all the code chunks included in the [Tutorial](https://juliagraphs.org/MultilayerGraphs.jl/dev/#Tutorial) section of the package development documentation that should be effectively run by the user and tested it runs smoothly on a new empty environment.

All the code chunks such as:

```julia
Node(id::String)
```

```julia
MultilayerVertex(
                node::Node,                            # The Node that the vertex will represent
                layer::Union{Nothing,Symbol},          # The layer which the `Node` will be represented in. Should be set to `nothing` when constructing layers.
                metadata::Union{<:NamedTuple,<:Tuple} # The metadata associated to this vertex
)
```

should not be run since their purpose is just to showcase the signature of structs' constructors (e.g. `Node`, `MultilayerVertex`, `Layer`, etc.).
