### 2023-02-28 #5116 @pitmonticone
## Author Checklist

This should be the complete list of the changes requested by the reviewers:

- [x] Change the [README](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/README.md), [example](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.md) and [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/) specifying the minimum version. The stable documentation will be updated with the registration of the post-publication release
- [x] Add [issue templates](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/main/.github/ISSUE_TEMPLATE)
- [x] Add the [environment configuration file](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/Project.toml) and the necessary code to activate and instantiate that environment in order to ensure full reproducibility both in [example.jl](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/example.jl) and [example.md](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/example/README.md).
- [x] Update documentation code to install all necessary dependencies in [dev documentation](https://juliagraphs.org/MultilayerGraphs.jl/dev/). The stable documentation will be updated with the registration of the post-publication release.
- [x] Rename `example.md` into `README.md`
- [x] Add a [script](https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/main/docs/src/tutorial.jl) containing all the code chunks included in the [Tutorial](https://juliagraphs.org/MultilayerGraphs.jl/dev/#Tutorial) section of the package development documentation that should be effectively run by the user to test it runs smoothly on a new empty environment

CC: @diehlpk @TomKellyGenetics @bgailleton @ClaudMor
