### 2023-03-01 #5116 @TomKellyGenetics
Thanks for your quick feedback @pitmonticone I was indeed unsure whether these were intended to be executed or serve to illustrate the general usage.

I have a minor issue which may be due to my environment (using Japanese locale settings and encoders).
> ≈ or ÷ characters are not recognised on my system

Is it possible to replace these with generic characters (isapprox and "/"" respectively) to enable users with other local settings to run it? I've tried to install a patched version from GitHub but the issue persists.

https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/e4a5ff228421ba686a0f15160c1766fe68b3a98e/docs/src/tutorial.jl#L33
```suggestion
-    _ne = rand(1:(_nv*(_nv-1)) ÷ 2 )
+    _ne = rand(1:(_nv*(_nv-1)) / 2 )
```
https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/e4a5ff228421ba686a0f15160c1766fe68b3a98e/src/subgraphs/layer.jl#L207
```suggestion
-     maxe = directed ? _nv * (_nv - 1) : _nv * (_nv - 1) ÷ 2
+     maxe = directed ? _nv * (_nv - 1) : _nv * (_nv - 1) / 2
```
https://github.com/JuliaGraphs/MultilayerGraphs.jl/blob/5d116429bbba1f078061f60d5d0ae229576abaa9/test/runtests.jl#L25
```suggestion
- _ne = rand(_nv:(((_nv * (_nv - 1)) ÷ 2) - 1))
+ _ne = rand(_nv:(((_nv * (_nv - 1)) / 2) - 1))
```

I'll try testing it on a different system in case this problem is specific to my settings. I don't think this is a serious issue with the package if I am able to run it elsewhere but I will share it in case it helps make it compatible with a wider range of potential users.
