### 2023-03-01 #5116 @pitmonticone
Hi @TomKellyGenetics,

Thank you very much for checking the wide applicability of our tutorial.

The Unicode symbol `≈` stands for the approximate equality function `isapprox()` while the symbol `÷` represents the floor division function `div()` (not the standard division `/`).

You might find it useful to try to substitute those symbols with the respective functions and check if it all works fine in your environment.
