### 2023-03-03 #5116 @TomKellyGenetics
I'm still having trouble running the examples but the issue seems to be specific to my environment (CentOS 7 server). I've tried installing a fork without these macros from GitHub but the problem persists. I'll try testing it with MacOS or a newer Linux distribution (on other systems or cloud-based VM) and file issues with the repository if I find any further problems if they are reproducible. If some older systems are not supported, please document these "system requirements" in the installation section.

I tried installing Julia 1.8 and MultilayerGraphs 1.1.3 on a fresh Windows 10 PC. I am able to run the `docs/src/tutorial.jl` and `test/runtests.jl` scripts as interactive sessions and non-interactive scripts without errors. As I am now able to confirm functionality of the software described in the paper, I will recommend acceptance of this manuscript.

Tanti auguri Claudio and Pietro, excellent work. You've developed high quality software suitable for publication in my opinion. I am impressed by how organised you are and handling reviewers comments in a timely manner.

Editors (@diehlpk) please note that both reviews have been completed.
