### 2023-03-03 JuliaGraphs/MultilayerGraphs.jl#129 @bgailleton
Thanks a lot for addressing these comments!
