### 2023-03-03 #5116 @pitmonticone
I've just:

- [ ] Generated a [new release](https://github.com/JuliaGraphs/MultilayerGraphs.jl/releases/tag/v1.1.4) including all changes during the review. The release number is: **1.1.4**.
- [ ] Uploaded the release to [Zenodo](https://doi.org/10.5281/zenodo.7009172). The universal Zenodo DOI is **10.5281/zenodo.7009172** while the v1.1.4-specific Zenodo DOI is **10.5281/zenodo.7697094**.

Thank you very much @TomKellyGenetics and @bgailleton for your helpful reviews and @diehlpk for your editorial management.

CC: @ClaudMor.
