### 2023-03-04 #5116 @tarleb
@danielskatz I have created a PR that should fix the warning: JuliaGraphs/MultilayerGraphs.jl#136
