### 2023-03-05 #5116 @pitmonticone
@danielskatz

Is it possible to
- change the citation string by ensuring that both authors (who have contributed equally) get mentioned (i.e. from "Moroni et al." to "Moroni and Monticone");
- change the BibTeX handle from "Moroni2023" to "MoroniMonticone2023"?

Thanks in advance.
