### 2023-03-07 #5116 @TomKellyGenetics
Quick update on testing the package.

I've confirmed this package runs as described on Mac OS (in addition to Windows). I tried running it on a different Linux system but there problems installing dependencies in my environment. This seems to be an issue with dependent packages, not the published work, therefore I think it will not be necessary to report issues to the developers. If that changes, I'll submit issues directly to the package GitHub repository separate from the JOSS review.

@pitmonticone @ClaudMor Thank you for your cooperation.
