### 2023-10-04 #5116 @pitmonticone
Hi @arfon @danielskatz,

Where can I find the .tex source of the paper?
