### 2023-10-04 #5116 @danielskatz
The source of the paper is https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/JOSS/paper
