### 2024-07-01 #5116 @pitmonticone
@danielskatz I have updated my affiliation in the [source file](https://github.com/JuliaGraphs/MultilayerGraphs.jl/tree/JOSS/paper). Is it possible to update it on the publication as well?
